Module aws_sdk_mediaconvert::error

Modules

bad_request_exception

See BadRequestException

conflict_exception

See ConflictException

forbidden_exception

See ForbiddenException

internal_server_error_exception

See InternalServerErrorException

not_found_exception

See NotFoundException

too_many_requests_exception

See TooManyRequestsException

Structs

AssociateCertificateError

BadRequestException

The service can’t process your request because of a problem in the request. Please check your request form and syntax.

CancelJobError

ConflictException

The service couldn’t complete your request because there is a conflict with the current state of the resource.

CreateJobError

CreateJobTemplateError

CreatePresetError

CreateQueueError

DeleteJobTemplateError

DeletePresetError

DeleteQueueError

DescribeEndpointsError

DisassociateCertificateError

ForbiddenException

You don’t have permissions for this action with the credentials you sent.

GetJobError

GetJobTemplateError

GetPresetError

GetQueueError

InternalServerErrorException

The service encountered an unexpected condition and can’t fulfill your request.

ListJobTemplatesError

ListJobsError

ListPresetsError

ListQueuesError

ListTagsForResourceError

NotFoundException

The resource you requested doesn’t exist.

TagResourceError

TooManyRequestsException

Too many requests have been sent in too short of a time. The service limits the rate at which it will accept requests.

UntagResourceError

UpdateJobTemplateError

UpdatePresetError

UpdateQueueError

Enums

AssociateCertificateErrorKind

CancelJobErrorKind

CreateJobErrorKind

CreateJobTemplateErrorKind

CreatePresetErrorKind

CreateQueueErrorKind

DeleteJobTemplateErrorKind

DeletePresetErrorKind

DeleteQueueErrorKind

DescribeEndpointsErrorKind

DisassociateCertificateErrorKind

GetJobErrorKind

GetJobTemplateErrorKind

GetPresetErrorKind

GetQueueErrorKind

ListJobTemplatesErrorKind

ListJobsErrorKind

ListPresetsErrorKind

ListQueuesErrorKind

ListTagsForResourceErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateJobTemplateErrorKind

UpdatePresetErrorKind

UpdateQueueErrorKind