Struct aws_sdk_mediaconvert::input::create_job_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for CreateJobInput

Implementations

impl Builder

pub fn acceleration_settings(self, input: AccelerationSettings) -> Self

Optional. Accelerated transcoding can significantly speed up jobs with long, visually complex content. Outputs that use this feature incur pro-tier pricing. For information about feature limitations, see the AWS Elemental MediaConvert User Guide.

pub fn set_acceleration_settings(
self,
input: Option<AccelerationSettings>
) -> Self

pub fn billing_tags_source(self, input: BillingTagsSource) -> Self

Optional. Choose a tag type that AWS Billing and Cost Management will use to sort your AWS Elemental MediaConvert costs on any billing report that you set up. Any transcoding outputs that don’t have an associated tag will appear in your billing report unsorted. If you don’t choose a valid value for this field, your job outputs will appear on the billing report unsorted.

pub fn set_billing_tags_source(self, input: Option<BillingTagsSource>) -> Self

pub fn client_request_token(self, input: impl Into<String>) -> Self

Optional. Idempotency token for CreateJob operation.

pub fn set_client_request_token(self, input: Option<String>) -> Self

pub fn hop_destinations(self, input: impl Into<HopDestination>) -> Self

pub fn set_hop_destinations(self, input: Option<Vec<HopDestination>>) -> Self

pub fn job_template(self, input: impl Into<String>) -> Self

Optional. When you create a job, you can either specify a job template or specify the transcoding settings individually.

pub fn set_job_template(self, input: Option<String>) -> Self

pub fn priority(self, input: i32) -> Self

Optional. Specify the relative priority for this job. In any given queue, the service begins processing the job with the highest value first. When more than one job has the same priority, the service begins processing the job that you submitted first. If you don’t specify a priority, the service uses the default value 0.

pub fn set_priority(self, input: Option<i32>) -> Self

pub fn queue(self, input: impl Into<String>) -> Self

Optional. When you create a job, you can specify a queue to send it to. If you don’t specify, the job will go to the default queue. For more about queues, see the User Guide topic at https://docs.aws.amazon.com/mediaconvert/latest/ug/what-is.html.

pub fn set_queue(self, input: Option<String>) -> Self

pub fn role(self, input: impl Into<String>) -> Self

Required. The IAM role you use for creating this job. For details about permissions, see the User Guide topic at the User Guide at https://docs.aws.amazon.com/mediaconvert/latest/ug/iam-role.html.

pub fn set_role(self, input: Option<String>) -> Self

pub fn settings(self, input: JobSettings) -> Self

JobSettings contains all the transcode settings for a job.

pub fn set_settings(self, input: Option<JobSettings>) -> Self

pub fn simulate_reserved_queue(self, input: SimulateReservedQueue) -> Self

Optional. Enable this setting when you run a test job to estimate how many reserved transcoding slots (RTS) you need. When this is enabled, MediaConvert runs your job from an on-demand queue with similar performance to what you will see with one RTS in a reserved queue. This setting is disabled by default.

pub fn set_simulate_reserved_queue(
self,
input: Option<SimulateReservedQueue>
) -> Self

pub fn status_update_interval(self, input: StatusUpdateInterval) -> Self

Optional. Specify how often MediaConvert sends STATUS_UPDATE events to Amazon CloudWatch Events. Set the interval, in seconds, between status updates. MediaConvert sends an update at this interval from the time the service begins processing your job to the time it completes the transcode or encounters an error.

pub fn set_status_update_interval(
self,
input: Option<StatusUpdateInterval>
) -> Self

pub fn tags(self, k: impl Into<String>, v: impl Into<String>) -> Self

pub fn set_tags(self, input: Option<HashMap<String, String>>) -> Self

pub fn user_metadata(self, k: impl Into<String>, v: impl Into<String>) -> Self

pub fn set_user_metadata(self, input: Option<HashMap<String, String>>) -> Self

pub fn build(self) -> Result<CreateJobInput, BuildError>

Consumes the builder and constructs a CreateJobInput