Struct aws_sdk_mediaconvert::model::ac3_settings::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for Ac3Settings

Implementations

impl Builder

pub fn bitrate(self, input: i32) -> Self

Specify the average bitrate in bits per second. Valid bitrates depend on the coding mode.

pub fn set_bitrate(self, input: Option<i32>) -> Self

pub fn bitstream_mode(self, input: Ac3BitstreamMode) -> Self

Specify the bitstream mode for the AC-3 stream that the encoder emits. For more information about the AC3 bitstream mode, see ATSC A/52-2012 (Annex E).

pub fn set_bitstream_mode(self, input: Option<Ac3BitstreamMode>) -> Self

pub fn coding_mode(self, input: Ac3CodingMode) -> Self

Dolby Digital coding mode. Determines number of channels.

pub fn set_coding_mode(self, input: Option<Ac3CodingMode>) -> Self

pub fn dialnorm(self, input: i32) -> Self

Sets the dialnorm for the output. If blank and input audio is Dolby Digital, dialnorm will be passed through.

pub fn set_dialnorm(self, input: Option<i32>) -> Self

pub fn dynamic_range_compression_line(
self,
input: Ac3DynamicRangeCompressionLine
) -> Self

Choose the Dolby Digital dynamic range control (DRC) profile that MediaConvert uses when encoding the metadata in the Dolby Digital stream for the line operating mode. Related setting: When you use this setting, MediaConvert ignores any value you provide for Dynamic range compression profile (DynamicRangeCompressionProfile). For information about the Dolby Digital DRC operating modes and profiles, see the Dynamic Range Control chapter of the Dolby Metadata Guide at https://developer.dolby.com/globalassets/professional/documents/dolby-metadata-guide.pdf.

pub fn set_dynamic_range_compression_line(
self,
input: Option<Ac3DynamicRangeCompressionLine>
) -> Self

pub fn dynamic_range_compression_profile(
self,
input: Ac3DynamicRangeCompressionProfile
) -> Self

When you want to add Dolby dynamic range compression (DRC) signaling to your output stream, we recommend that you use the mode-specific settings instead of Dynamic range compression profile (DynamicRangeCompressionProfile). The mode-specific settings are Dynamic range compression profile, line mode (dynamicRangeCompressionLine) and Dynamic range compression profile, RF mode (dynamicRangeCompressionRf). Note that when you specify values for all three settings, MediaConvert ignores the value of this setting in favor of the mode-specific settings. If you do use this setting instead of the mode-specific settings, choose None (NONE) to leave out DRC signaling. Keep the default Film standard (FILM_STANDARD) to set the profile to Dolby’s film standard profile for all operating modes.

pub fn set_dynamic_range_compression_profile(
self,
input: Option<Ac3DynamicRangeCompressionProfile>
) -> Self

pub fn dynamic_range_compression_rf(
self,
input: Ac3DynamicRangeCompressionRf
) -> Self

Choose the Dolby Digital dynamic range control (DRC) profile that MediaConvert uses when encoding the metadata in the Dolby Digital stream for the RF operating mode. Related setting: When you use this setting, MediaConvert ignores any value you provide for Dynamic range compression profile (DynamicRangeCompressionProfile). For information about the Dolby Digital DRC operating modes and profiles, see the Dynamic Range Control chapter of the Dolby Metadata Guide at https://developer.dolby.com/globalassets/professional/documents/dolby-metadata-guide.pdf.

pub fn set_dynamic_range_compression_rf(
self,
input: Option<Ac3DynamicRangeCompressionRf>
) -> Self

pub fn lfe_filter(self, input: Ac3LfeFilter) -> Self

Applies a 120Hz lowpass filter to the LFE channel prior to encoding. Only valid with 3_2_LFE coding mode.

pub fn set_lfe_filter(self, input: Option<Ac3LfeFilter>) -> Self

pub fn metadata_control(self, input: Ac3MetadataControl) -> Self

When set to FOLLOW_INPUT, encoder metadata will be sourced from the DD, DD+, or DolbyE decoder that supplied this audio data. If audio was not supplied from one of these streams, then the static metadata settings will be used.

pub fn set_metadata_control(self, input: Option<Ac3MetadataControl>) -> Self

pub fn sample_rate(self, input: i32) -> Self

This value is always 48000. It represents the sample rate in Hz.

pub fn set_sample_rate(self, input: Option<i32>) -> Self

pub fn build(self) -> Ac3Settings

Consumes the builder and constructs a Ac3Settings