Struct aws_sdk_mediaconvert::model::av1_settings::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for Av1Settings

Implementations

impl Builder

pub fn adaptive_quantization(self, input: Av1AdaptiveQuantization) -> Self

Specify the strength of any adaptive quantization filters that you enable. The value that you choose here applies to Spatial adaptive quantization (spatialAdaptiveQuantization).

pub fn set_adaptive_quantization(
self,
input: Option<Av1AdaptiveQuantization>
) -> Self

pub fn framerate_control(self, input: Av1FramerateControl) -> Self

If you are using the console, use the Framerate setting to specify the frame rate for this output. If you want to keep the same frame rate as the input video, choose Follow source. If you want to do frame rate conversion, choose a frame rate from the dropdown list or choose Custom. The framerates shown in the dropdown list are decimal approximations of fractions. If you choose Custom, specify your frame rate as a fraction. If you are creating your transcoding job specification as a JSON file without the console, use FramerateControl to specify which value the service uses for the frame rate for this output. Choose INITIALIZE_FROM_SOURCE if you want the service to use the frame rate from the input. Choose SPECIFIED if you want the service to use the frame rate you specify in the settings FramerateNumerator and FramerateDenominator.

pub fn set_framerate_control(self, input: Option<Av1FramerateControl>) -> Self

pub fn framerate_conversion_algorithm(
self,
input: Av1FramerateConversionAlgorithm
) -> Self

Choose the method that you want MediaConvert to use when increasing or decreasing the frame rate. We recommend using drop duplicate (DUPLICATE_DROP) for numerically simple conversions, such as 60 fps to 30 fps. For numerically complex conversions, you can use interpolate (INTERPOLATE) to avoid stutter. This results in a smooth picture, but might introduce undesirable video artifacts. For complex frame rate conversions, especially if your source video has already been converted from its original cadence, use FrameFormer (FRAMEFORMER) to do motion-compensated interpolation. FrameFormer chooses the best conversion method frame by frame. Note that using FrameFormer increases the transcoding time and incurs a significant add-on cost.

pub fn set_framerate_conversion_algorithm(
self,
input: Option<Av1FramerateConversionAlgorithm>
) -> Self

pub fn framerate_denominator(self, input: i32) -> Self

When you use the API for transcode jobs that use frame rate conversion, specify the frame rate as a fraction. For example, 24000 / 1001 = 23.976 fps. Use FramerateDenominator to specify the denominator of this fraction. In this example, use 1001 for the value of FramerateDenominator. When you use the console for transcode jobs that use frame rate conversion, provide the value as a decimal number for Framerate. In this example, specify 23.976.

pub fn set_framerate_denominator(self, input: Option<i32>) -> Self

pub fn framerate_numerator(self, input: i32) -> Self

When you use the API for transcode jobs that use frame rate conversion, specify the frame rate as a fraction. For example, 24000 / 1001 = 23.976 fps. Use FramerateNumerator to specify the numerator of this fraction. In this example, use 24000 for the value of FramerateNumerator. When you use the console for transcode jobs that use frame rate conversion, provide the value as a decimal number for Framerate. In this example, specify 23.976.

pub fn set_framerate_numerator(self, input: Option<i32>) -> Self

pub fn gop_size(self, input: f64) -> Self

Specify the GOP length (keyframe interval) in frames. With AV1, MediaConvert doesn’t support GOP length in seconds. This value must be greater than zero and preferably equal to 1 + ((numberBFrames + 1) * x), where x is an integer value.

pub fn set_gop_size(self, input: Option<f64>) -> Self

pub fn max_bitrate(self, input: i32) -> Self

Maximum bitrate in bits/second. For example, enter five megabits per second as 5000000. Required when Rate control mode is QVBR.

pub fn set_max_bitrate(self, input: Option<i32>) -> Self

pub fn number_b_frames_between_reference_frames(self, input: i32) -> Self

Specify from the number of B-frames, in the range of 0-15. For AV1 encoding, we recommend using 7 or 15. Choose a larger number for a lower bitrate and smaller file size; choose a smaller number for better video quality.

pub fn set_number_b_frames_between_reference_frames(
self,
input: Option<i32>
) -> Self

pub fn qvbr_settings(self, input: Av1QvbrSettings) -> Self

Settings for quality-defined variable bitrate encoding with the H.265 codec. Use these settings only when you set QVBR for Rate control mode (RateControlMode).

pub fn set_qvbr_settings(self, input: Option<Av1QvbrSettings>) -> Self

pub fn rate_control_mode(self, input: Av1RateControlMode) -> Self

‘With AV1 outputs, for rate control mode, MediaConvert supports only quality-defined variable bitrate (QVBR). You can’‘t use CBR or VBR.’

pub fn set_rate_control_mode(self, input: Option<Av1RateControlMode>) -> Self

pub fn slices(self, input: i32) -> Self

Specify the number of slices per picture. This value must be 1, 2, 4, 8, 16, or 32. For progressive pictures, this value must be less than or equal to the number of macroblock rows. For interlaced pictures, this value must be less than or equal to half the number of macroblock rows.

pub fn set_slices(self, input: Option<i32>) -> Self

pub fn spatial_adaptive_quantization(
self,
input: Av1SpatialAdaptiveQuantization
) -> Self

Keep the default value, Enabled (ENABLED), to adjust quantization within each frame based on spatial variation of content complexity. When you enable this feature, the encoder uses fewer bits on areas that can sustain more distortion with no noticeable visual degradation and uses more bits on areas where any small distortion will be noticeable. For example, complex textured blocks are encoded with fewer bits and smooth textured blocks are encoded with more bits. Enabling this feature will almost always improve your video quality. Note, though, that this feature doesn’t take into account where the viewer’s attention is likely to be. If viewers are likely to be focusing their attention on a part of the screen with a lot of complex texture, you might choose to disable this feature. Related setting: When you enable spatial adaptive quantization, set the value for Adaptive quantization (adaptiveQuantization) depending on your content. For homogeneous content, such as cartoons and video games, set it to Low. For content with a wider variety of textures, set it to High or Higher.

pub fn set_spatial_adaptive_quantization(
self,
input: Option<Av1SpatialAdaptiveQuantization>
) -> Self

pub fn build(self) -> Av1Settings

Consumes the builder and constructs a Av1Settings