Struct aws_sdk_mediaconvert::model::hls_group_settings::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for HlsGroupSettings

Implementations

impl Builder

pub fn ad_markers(self, input: impl Into<HlsAdMarkers>) -> Self

pub fn set_ad_markers(self, input: Option<Vec<HlsAdMarkers>>) -> Self

pub fn additional_manifests(
self,
input: impl Into<HlsAdditionalManifest>
) -> Self

pub fn set_additional_manifests(
self,
input: Option<Vec<HlsAdditionalManifest>>
) -> Self

pub fn audio_only_header(self, input: HlsAudioOnlyHeader) -> Self

Ignore this setting unless you are using FairPlay DRM with Verimatrix and you encounter playback issues. Keep the default value, Include (INCLUDE), to output audio-only headers. Choose Exclude (EXCLUDE) to remove the audio-only headers from your audio segments.

pub fn set_audio_only_header(self, input: Option<HlsAudioOnlyHeader>) -> Self

pub fn base_url(self, input: impl Into<String>) -> Self

A partial URI prefix that will be prepended to each output in the media .m3u8 file. Can be used if base manifest is delivered from a different URL than the main .m3u8 file.

pub fn set_base_url(self, input: Option<String>) -> Self

pub fn caption_language_mappings(
self,
input: impl Into<HlsCaptionLanguageMapping>
) -> Self

pub fn set_caption_language_mappings(
self,
input: Option<Vec<HlsCaptionLanguageMapping>>
) -> Self

pub fn caption_language_setting(self, input: HlsCaptionLanguageSetting) -> Self

Applies only to 608 Embedded output captions. Insert: Include CLOSED-CAPTIONS lines in the manifest. Specify at least one language in the CC1 Language Code field. One CLOSED-CAPTION line is added for each Language Code you specify. Make sure to specify the languages in the order in which they appear in the original source (if the source is embedded format) or the order of the caption selectors (if the source is other than embedded). Otherwise, languages in the manifest will not match up properly with the output captions. None: Include CLOSED-CAPTIONS=NONE line in the manifest. Omit: Omit any CLOSED-CAPTIONS line from the manifest.

pub fn set_caption_language_setting(
self,
input: Option<HlsCaptionLanguageSetting>
) -> Self

pub fn client_cache(self, input: HlsClientCache) -> Self

Disable this setting only when your workflow requires the #EXT-X-ALLOW-CACHE:no tag. Otherwise, keep the default value Enabled (ENABLED) and control caching in your video distribution set up. For example, use the Cache-Control http header.

pub fn set_client_cache(self, input: Option<HlsClientCache>) -> Self

pub fn codec_specification(self, input: HlsCodecSpecification) -> Self

Specification to use (RFC-6381 or the default RFC-4281) during m3u8 playlist generation.

pub fn set_codec_specification(
self,
input: Option<HlsCodecSpecification>
) -> Self

pub fn destination(self, input: impl Into<String>) -> Self

Use Destination (Destination) to specify the S3 output location and the output filename base. Destination accepts format identifiers. If you do not specify the base filename in the URI, the service will use the filename of the input file. If your job has multiple inputs, the service uses the filename of the first input file.

pub fn set_destination(self, input: Option<String>) -> Self

pub fn destination_settings(self, input: DestinationSettings) -> Self

Settings associated with the destination. Will vary based on the type of destination

pub fn set_destination_settings(
self,
input: Option<DestinationSettings>
) -> Self

pub fn directory_structure(self, input: HlsDirectoryStructure) -> Self

Indicates whether segments should be placed in subdirectories.

pub fn set_directory_structure(
self,
input: Option<HlsDirectoryStructure>
) -> Self

pub fn encryption(self, input: HlsEncryptionSettings) -> Self

DRM settings.

pub fn set_encryption(self, input: Option<HlsEncryptionSettings>) -> Self

pub fn image_based_trick_play(self, input: HlsImageBasedTrickPlay) -> Self

Specify whether MediaConvert generates images for trick play. Keep the default value, None (NONE), to not generate any images. Choose Thumbnail (THUMBNAIL) to generate tiled thumbnails. Choose Thumbnail and full frame (THUMBNAIL_AND_FULLFRAME) to generate tiled thumbnails and full-resolution images of single frames. MediaConvert creates a child manifest for each set of images that you generate and adds corresponding entries to the parent manifest. A common application for these images is Roku trick mode. The thumbnails and full-frame images that MediaConvert creates with this feature are compatible with this Roku specification: https://developer.roku.com/docs/developer-program/media-playback/trick-mode/hls-and-dash.md

pub fn set_image_based_trick_play(
self,
input: Option<HlsImageBasedTrickPlay>
) -> Self

pub fn manifest_compression(self, input: HlsManifestCompression) -> Self

When set to GZIP, compresses HLS playlist.

pub fn set_manifest_compression(
self,
input: Option<HlsManifestCompression>
) -> Self

pub fn manifest_duration_format(self, input: HlsManifestDurationFormat) -> Self

Indicates whether the output manifest should use floating point values for segment duration.

pub fn set_manifest_duration_format(
self,
input: Option<HlsManifestDurationFormat>
) -> Self

pub fn min_final_segment_length(self, input: f64) -> Self

Keep this setting at the default value of 0, unless you are troubleshooting a problem with how devices play back the end of your video asset. If you know that player devices are hanging on the final segment of your video because the length of your final segment is too short, use this setting to specify a minimum final segment length, in seconds. Choose a value that is greater than or equal to 1 and less than your segment length. When you specify a value for this setting, the encoder will combine any final segment that is shorter than the length that you specify with the previous segment. For example, your segment length is 3 seconds and your final segment is .5 seconds without a minimum final segment length; when you set the minimum final segment length to 1, your final segment is 3.5 seconds.

pub fn set_min_final_segment_length(self, input: Option<f64>) -> Self

pub fn min_segment_length(self, input: i32) -> Self

When set, Minimum Segment Size is enforced by looking ahead and back within the specified range for a nearby avail and extending the segment size if needed.

pub fn set_min_segment_length(self, input: Option<i32>) -> Self

pub fn output_selection(self, input: HlsOutputSelection) -> Self

Indicates whether the .m3u8 manifest file should be generated for this HLS output group.

pub fn set_output_selection(self, input: Option<HlsOutputSelection>) -> Self

pub fn program_date_time(self, input: HlsProgramDateTime) -> Self

Includes or excludes EXT-X-PROGRAM-DATE-TIME tag in .m3u8 manifest files. The value is calculated as follows: either the program date and time are initialized using the input timecode source, or the time is initialized using the input timecode source and the date is initialized using the timestamp_offset.

pub fn set_program_date_time(self, input: Option<HlsProgramDateTime>) -> Self

pub fn program_date_time_period(self, input: i32) -> Self

Period of insertion of EXT-X-PROGRAM-DATE-TIME entry, in seconds.

pub fn set_program_date_time_period(self, input: Option<i32>) -> Self

pub fn segment_control(self, input: HlsSegmentControl) -> Self

When set to SINGLE_FILE, emits program as a single media resource (.ts) file, uses #EXT-X-BYTERANGE tags to index segment for playback.

pub fn set_segment_control(self, input: Option<HlsSegmentControl>) -> Self

pub fn segment_length(self, input: i32) -> Self

Specify the length, in whole seconds, of each segment. When you don’t specify a value, MediaConvert defaults to 10. Related settings: Use Segment length control (SegmentLengthControl) to specify whether the encoder enforces this value strictly. Use Segment control (HlsSegmentControl) to specify whether MediaConvert creates separate segment files or one content file that has metadata to mark the segment boundaries.

pub fn set_segment_length(self, input: Option<i32>) -> Self

pub fn segment_length_control(self, input: HlsSegmentLengthControl) -> Self

Specify how you want MediaConvert to determine the segment length. Choose Exact (EXACT) to have the encoder use the exact length that you specify with the setting Segment length (SegmentLength). This might result in extra I-frames. Choose Multiple of GOP (GOP_MULTIPLE) to have the encoder round up the segment lengths to match the next GOP boundary.

pub fn set_segment_length_control(
self,
input: Option<HlsSegmentLengthControl>
) -> Self

pub fn segments_per_subdirectory(self, input: i32) -> Self

Number of segments to write to a subdirectory before starting a new one. directoryStructure must be SINGLE_DIRECTORY for this setting to have an effect.

pub fn set_segments_per_subdirectory(self, input: Option<i32>) -> Self

pub fn stream_inf_resolution(self, input: HlsStreamInfResolution) -> Self

Include or exclude RESOLUTION attribute for video in EXT-X-STREAM-INF tag of variant manifest.

pub fn set_stream_inf_resolution(
self,
input: Option<HlsStreamInfResolution>
) -> Self

pub fn target_duration_compatibility_mode(
self,
input: HlsTargetDurationCompatibilityMode
) -> Self

When set to LEGACY, the segment target duration is always rounded up to the nearest integer value above its current value in seconds. When set to SPEC\_COMPLIANT, the segment target duration is rounded up to the nearest integer value if fraction seconds are greater than or equal to 0.5 (>= 0.5) and rounded down if less than 0.5 (< 0.5). You may need to use LEGACY if your client needs to ensure that the target duration is always longer than the actual duration of the segment. Some older players may experience interrupted playback when the actual duration of a track in a segment is longer than the target duration.

pub fn set_target_duration_compatibility_mode(
self,
input: Option<HlsTargetDurationCompatibilityMode>
) -> Self

pub fn timed_metadata_id3_frame(self, input: HlsTimedMetadataId3Frame) -> Self

Indicates ID3 frame that has the timecode.

pub fn set_timed_metadata_id3_frame(
self,
input: Option<HlsTimedMetadataId3Frame>
) -> Self

pub fn timed_metadata_id3_period(self, input: i32) -> Self

Timed Metadata interval in seconds.

pub fn set_timed_metadata_id3_period(self, input: Option<i32>) -> Self

pub fn timestamp_delta_milliseconds(self, input: i32) -> Self

Provides an extra millisecond delta offset to fine tune the timestamps.

pub fn set_timestamp_delta_milliseconds(self, input: Option<i32>) -> Self

pub fn build(self) -> HlsGroupSettings

Consumes the builder and constructs a HlsGroupSettings