Struct aws_sdk_mediaconvert::model::input_template::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for InputTemplate

Implementations

impl Builder

pub fn audio_selector_groups(
self,
k: impl Into<String>,
v: impl Into<AudioSelectorGroup>
) -> Self

pub fn set_audio_selector_groups(
self,
input: Option<HashMap<String, AudioSelectorGroup>>
) -> Self

pub fn audio_selectors(
self,
k: impl Into<String>,
v: impl Into<AudioSelector>
) -> Self

pub fn set_audio_selectors(
self,
input: Option<HashMap<String, AudioSelector>>
) -> Self

pub fn caption_selectors(
self,
k: impl Into<String>,
v: impl Into<CaptionSelector>
) -> Self

pub fn set_caption_selectors(
self,
input: Option<HashMap<String, CaptionSelector>>
) -> Self

pub fn crop(self, input: Rectangle) -> Self

Use Cropping selection (crop) to specify the video area that the service will include in the output video frame. If you specify a value here, it will override any value that you specify in the output setting Cropping selection (crop).

pub fn set_crop(self, input: Option<Rectangle>) -> Self

pub fn deblock_filter(self, input: InputDeblockFilter) -> Self

Enable Deblock (InputDeblockFilter) to produce smoother motion in the output. Default is disabled. Only manually controllable for MPEG2 and uncompressed video inputs.

pub fn set_deblock_filter(self, input: Option<InputDeblockFilter>) -> Self

pub fn denoise_filter(self, input: InputDenoiseFilter) -> Self

Enable Denoise (InputDenoiseFilter) to filter noise from the input. Default is disabled. Only applicable to MPEG2, H.264, H.265, and uncompressed video inputs.

pub fn set_denoise_filter(self, input: Option<InputDenoiseFilter>) -> Self

pub fn filter_enable(self, input: InputFilterEnable) -> Self

Specify how the transcoding service applies the denoise and deblock filters. You must also enable the filters separately, with Denoise (InputDenoiseFilter) and Deblock (InputDeblockFilter). * Auto - The transcoding service determines whether to apply filtering, depending on input type and quality. * Disable - The input is not filtered. This is true even if you use the API to enable them in (InputDeblockFilter) and (InputDeblockFilter). * Force - The input is filtered regardless of input type.

pub fn set_filter_enable(self, input: Option<InputFilterEnable>) -> Self

pub fn filter_strength(self, input: i32) -> Self

Use Filter strength (FilterStrength) to adjust the magnitude the input filter settings (Deblock and Denoise). The range is -5 to 5. Default is 0.

pub fn set_filter_strength(self, input: Option<i32>) -> Self

pub fn image_inserter(self, input: ImageInserter) -> Self

Enable the image inserter feature to include a graphic overlay on your video. Enable or disable this feature for each input individually. This setting is disabled by default.

pub fn set_image_inserter(self, input: Option<ImageInserter>) -> Self

pub fn input_clippings(self, input: impl Into<InputClipping>) -> Self

pub fn set_input_clippings(self, input: Option<Vec<InputClipping>>) -> Self

pub fn input_scan_type(self, input: InputScanType) -> Self

When you have a progressive segmented frame (PsF) input, use this setting to flag the input as PsF. MediaConvert doesn’t automatically detect PsF. Therefore, flagging your input as PsF results in better preservation of video quality when you do deinterlacing and frame rate conversion. If you don’t specify, the default value is Auto (AUTO). Auto is the correct setting for all inputs that are not PsF. Don’t set this value to PsF when your input is interlaced. Doing so creates horizontal interlacing artifacts.

pub fn set_input_scan_type(self, input: Option<InputScanType>) -> Self

pub fn position(self, input: Rectangle) -> Self

Use Selection placement (position) to define the video area in your output frame. The area outside of the rectangle that you specify here is black. If you specify a value here, it will override any value that you specify in the output setting Selection placement (position). If you specify a value here, this will override any AFD values in your input, even if you set Respond to AFD (RespondToAfd) to Respond (RESPOND). If you specify a value here, this will ignore anything that you specify for the setting Scaling Behavior (scalingBehavior).

pub fn set_position(self, input: Option<Rectangle>) -> Self

pub fn program_number(self, input: i32) -> Self

Use Program (programNumber) to select a specific program from within a multi-program transport stream. Note that Quad 4K is not currently supported. Default is the first program within the transport stream. If the program you specify doesn’t exist, the transcoding service will use this default.

pub fn set_program_number(self, input: Option<i32>) -> Self

pub fn psi_control(self, input: InputPsiControl) -> Self

Set PSI control (InputPsiControl) for transport stream inputs to specify which data the demux process to scans. * Ignore PSI - Scan all PIDs for audio and video. * Use PSI - Scan only PSI data.

pub fn set_psi_control(self, input: Option<InputPsiControl>) -> Self

pub fn timecode_source(self, input: InputTimecodeSource) -> Self

Use this Timecode source setting, located under the input settings (InputTimecodeSource), to specify how the service counts input video frames. This input frame count affects only the behavior of features that apply to a single input at a time, such as input clipping and synchronizing some captions formats. Choose Embedded (EMBEDDED) to use the timecodes in your input video. Choose Start at zero (ZEROBASED) to start the first frame at zero. Choose Specified start (SPECIFIEDSTART) to start the first frame at the timecode that you specify in the setting Start timecode (timecodeStart). If you don’t specify a value for Timecode source, the service will use Embedded by default. For more information about timecodes, see https://docs.aws.amazon.com/console/mediaconvert/timecode.

pub fn set_timecode_source(self, input: Option<InputTimecodeSource>) -> Self

pub fn timecode_start(self, input: impl Into<String>) -> Self

Specify the timecode that you want the service to use for this input’s initial frame. To use this setting, you must set the Timecode source setting, located under the input settings (InputTimecodeSource), to Specified start (SPECIFIEDSTART). For more information about timecodes, see https://docs.aws.amazon.com/console/mediaconvert/timecode.

pub fn set_timecode_start(self, input: Option<String>) -> Self

pub fn video_selector(self, input: VideoSelector) -> Self

Input video selectors contain the video settings for the input. Each of your inputs can have up to one video selector.

pub fn set_video_selector(self, input: Option<VideoSelector>) -> Self

pub fn build(self) -> InputTemplate

Consumes the builder and constructs a InputTemplate