Struct aws_sdk_mediaconvert::model::job::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for Job

Implementations

impl Builder

pub fn acceleration_settings(self, input: AccelerationSettings) -> Self

Accelerated transcoding can significantly speed up jobs with long, visually complex content.

pub fn set_acceleration_settings(
self,
input: Option<AccelerationSettings>
) -> Self

pub fn acceleration_status(self, input: AccelerationStatus) -> Self

Describes whether the current job is running with accelerated transcoding. For jobs that have Acceleration (AccelerationMode) set to DISABLED, AccelerationStatus is always NOT_APPLICABLE. For jobs that have Acceleration (AccelerationMode) set to ENABLED or PREFERRED, AccelerationStatus is one of the other states. AccelerationStatus is IN_PROGRESS initially, while the service determines whether the input files and job settings are compatible with accelerated transcoding. If they are, AcclerationStatus is ACCELERATED. If your input files and job settings aren’t compatible with accelerated transcoding, the service either fails your job or runs it without accelerated transcoding, depending on how you set Acceleration (AccelerationMode). When the service runs your job without accelerated transcoding, AccelerationStatus is NOT_ACCELERATED.

pub fn set_acceleration_status(self, input: Option<AccelerationStatus>) -> Self

pub fn arn(self, input: impl Into<String>) -> Self

An identifier for this resource that is unique within all of AWS.

pub fn set_arn(self, input: Option<String>) -> Self

pub fn billing_tags_source(self, input: BillingTagsSource) -> Self

The tag type that AWS Billing and Cost Management will use to sort your AWS Elemental MediaConvert costs on any billing report that you set up.

pub fn set_billing_tags_source(self, input: Option<BillingTagsSource>) -> Self

pub fn created_at(self, input: Instant) -> Self

The time, in Unix epoch format in seconds, when the job got created.

pub fn set_created_at(self, input: Option<Instant>) -> Self

pub fn current_phase(self, input: JobPhase) -> Self

A job’s phase can be PROBING, TRANSCODING OR UPLOADING

pub fn set_current_phase(self, input: Option<JobPhase>) -> Self

pub fn error_code(self, input: i32) -> Self

Error code for the job

pub fn set_error_code(self, input: Option<i32>) -> Self

pub fn error_message(self, input: impl Into<String>) -> Self

Error message of Job

pub fn set_error_message(self, input: Option<String>) -> Self

pub fn hop_destinations(self, input: impl Into<HopDestination>) -> Self

pub fn set_hop_destinations(self, input: Option<Vec<HopDestination>>) -> Self

pub fn id(self, input: impl Into<String>) -> Self

A portion of the job’s ARN, unique within your AWS Elemental MediaConvert resources

pub fn set_id(self, input: Option<String>) -> Self

pub fn job_percent_complete(self, input: i32) -> Self

An estimate of how far your job has progressed. This estimate is shown as a percentage of the total time from when your job leaves its queue to when your output files appear in your output Amazon S3 bucket. AWS Elemental MediaConvert provides jobPercentComplete in CloudWatch STATUS_UPDATE events and in the response to GetJob and ListJobs requests. The jobPercentComplete estimate is reliable for the following input containers: Quicktime, Transport Stream, MP4, and MXF. For some jobs, the service can’t provide information about job progress. In those cases, jobPercentComplete returns a null value.

pub fn set_job_percent_complete(self, input: Option<i32>) -> Self

pub fn job_template(self, input: impl Into<String>) -> Self

The job template that the job is created from, if it is created from a job template.

pub fn set_job_template(self, input: Option<String>) -> Self

pub fn messages(self, input: JobMessages) -> Self

Provides messages from the service about jobs that you have already successfully submitted.

pub fn set_messages(self, input: Option<JobMessages>) -> Self

pub fn output_group_details(self, input: impl Into<OutputGroupDetail>) -> Self

pub fn set_output_group_details(
self,
input: Option<Vec<OutputGroupDetail>>
) -> Self

pub fn priority(self, input: i32) -> Self

Relative priority on the job.

pub fn set_priority(self, input: Option<i32>) -> Self

pub fn queue(self, input: impl Into<String>) -> Self

When you create a job, you can specify a queue to send it to. If you don’t specify, the job will go to the default queue. For more about queues, see the User Guide topic at https://docs.aws.amazon.com/mediaconvert/latest/ug/what-is.html

pub fn set_queue(self, input: Option<String>) -> Self

pub fn queue_transitions(self, input: impl Into<QueueTransition>) -> Self

pub fn set_queue_transitions(self, input: Option<Vec<QueueTransition>>) -> Self

pub fn retry_count(self, input: i32) -> Self

The number of times that the service automatically attempted to process your job after encountering an error.

pub fn set_retry_count(self, input: Option<i32>) -> Self

pub fn role(self, input: impl Into<String>) -> Self

The IAM role you use for creating this job. For details about permissions, see the User Guide topic at the User Guide at https://docs.aws.amazon.com/mediaconvert/latest/ug/iam-role.html

pub fn set_role(self, input: Option<String>) -> Self

pub fn settings(self, input: JobSettings) -> Self

JobSettings contains all the transcode settings for a job.

pub fn set_settings(self, input: Option<JobSettings>) -> Self

pub fn simulate_reserved_queue(self, input: SimulateReservedQueue) -> Self

Enable this setting when you run a test job to estimate how many reserved transcoding slots (RTS) you need. When this is enabled, MediaConvert runs your job from an on-demand queue with similar performance to what you will see with one RTS in a reserved queue. This setting is disabled by default.

pub fn set_simulate_reserved_queue(
self,
input: Option<SimulateReservedQueue>
) -> Self

pub fn status(self, input: JobStatus) -> Self

A job’s status can be SUBMITTED, PROGRESSING, COMPLETE, CANCELED, or ERROR.

pub fn set_status(self, input: Option<JobStatus>) -> Self

pub fn status_update_interval(self, input: StatusUpdateInterval) -> Self

Specify how often MediaConvert sends STATUS_UPDATE events to Amazon CloudWatch Events. Set the interval, in seconds, between status updates. MediaConvert sends an update at this interval from the time the service begins processing your job to the time it completes the transcode or encounters an error.

pub fn set_status_update_interval(
self,
input: Option<StatusUpdateInterval>
) -> Self

pub fn timing(self, input: Timing) -> Self

Information about when jobs are submitted, started, and finished is specified in Unix epoch format in seconds.

pub fn set_timing(self, input: Option<Timing>) -> Self

pub fn user_metadata(self, k: impl Into<String>, v: impl Into<String>) -> Self

pub fn set_user_metadata(self, input: Option<HashMap<String, String>>) -> Self

pub fn build(self) -> Job

Consumes the builder and constructs a Job