Struct aws_sdk_mediaconvert::model::job_settings::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for JobSettings

Implementations

impl Builder

pub fn ad_avail_offset(self, input: i32) -> Self

When specified, this offset (in milliseconds) is added to the input Ad Avail PTS time.

pub fn set_ad_avail_offset(self, input: Option<i32>) -> Self

pub fn avail_blanking(self, input: AvailBlanking) -> Self

Settings for ad avail blanking. Video can be blanked or overlaid with an image, and audio muted during SCTE-35 triggered ad avails.

pub fn set_avail_blanking(self, input: Option<AvailBlanking>) -> Self

pub fn esam(self, input: EsamSettings) -> Self

Settings for Event Signaling And Messaging (ESAM). If you don’t do ad insertion, you can ignore these settings.

pub fn set_esam(self, input: Option<EsamSettings>) -> Self

pub fn extended_data_services(self, input: ExtendedDataServices) -> Self

Hexadecimal value as per EIA-608 Line 21 Data Services, section 9.5.1.5 05h Content Advisory.

pub fn set_extended_data_services(
self,
input: Option<ExtendedDataServices>
) -> Self

pub fn inputs(self, input: impl Into<Input>) -> Self

pub fn set_inputs(self, input: Option<Vec<Input>>) -> Self

pub fn kantar_watermark(self, input: KantarWatermarkSettings) -> Self

Use these settings only when you use Kantar watermarking. Specify the values that MediaConvert uses to generate and place Kantar watermarks in your output audio. These settings apply to every output in your job. In addition to specifying these values, you also need to store your Kantar credentials in AWS Secrets Manager. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/kantar-watermarking.html.

pub fn set_kantar_watermark(
self,
input: Option<KantarWatermarkSettings>
) -> Self

pub fn motion_image_inserter(self, input: MotionImageInserter) -> Self

Overlay motion graphics on top of your video. The motion graphics that you specify here appear on all outputs in all output groups. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/motion-graphic-overlay.html.

pub fn set_motion_image_inserter(
self,
input: Option<MotionImageInserter>
) -> Self

pub fn nielsen_configuration(self, input: NielsenConfiguration) -> Self

Settings for your Nielsen configuration. If you don’t do Nielsen measurement and analytics, ignore these settings. When you enable Nielsen configuration (nielsenConfiguration), MediaConvert enables PCM to ID3 tagging for all outputs in the job. To enable Nielsen configuration programmatically, include an instance of nielsenConfiguration in your JSON job specification. Even if you don’t include any children of nielsenConfiguration, you still enable the setting.

pub fn set_nielsen_configuration(
self,
input: Option<NielsenConfiguration>
) -> Self

pub fn nielsen_non_linear_watermark(
self,
input: NielsenNonLinearWatermarkSettings
) -> Self

Ignore these settings unless you are using Nielsen non-linear watermarking. Specify the values that MediaConvert uses to generate and place Nielsen watermarks in your output audio. In addition to specifying these values, you also need to set up your cloud TIC server. These settings apply to every output in your job. The MediaConvert implementation is currently with the following Nielsen versions: Nielsen Watermark SDK Version 5.2.1 Nielsen NLM Watermark Engine Version 1.2.7 Nielsen Watermark Authenticator [SID_TIC] Version [5.0.0]

pub fn set_nielsen_non_linear_watermark(
self,
input: Option<NielsenNonLinearWatermarkSettings>
) -> Self

pub fn output_groups(self, input: impl Into<OutputGroup>) -> Self

pub fn set_output_groups(self, input: Option<Vec<OutputGroup>>) -> Self

pub fn timecode_config(self, input: TimecodeConfig) -> Self

These settings control how the service handles timecodes throughout the job. These settings don’t affect input clipping.

pub fn set_timecode_config(self, input: Option<TimecodeConfig>) -> Self

pub fn timed_metadata_insertion(self, input: TimedMetadataInsertion) -> Self

Enable Timed metadata insertion (TimedMetadataInsertion) to include ID3 tags in any HLS outputs. To include timed metadata, you must enable it here, enable it in each output container, and specify tags and timecodes in ID3 insertion (Id3Insertion) objects.

pub fn set_timed_metadata_insertion(
self,
input: Option<TimedMetadataInsertion>
) -> Self

pub fn build(self) -> JobSettings

Consumes the builder and constructs a JobSettings