Struct aws_sdk_mediaconvert::model::m2ts_settings::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for M2tsSettings

Implementations

impl Builder

pub fn audio_buffer_model(self, input: M2tsAudioBufferModel) -> Self

Selects between the DVB and ATSC buffer models for Dolby Digital audio.

pub fn set_audio_buffer_model(self, input: Option<M2tsAudioBufferModel>) -> Self

pub fn audio_duration(self, input: M2tsAudioDuration) -> Self

Specify this setting only when your output will be consumed by a downstream repackaging workflow that is sensitive to very small duration differences between video and audio. For this situation, choose Match video duration (MATCH_VIDEO_DURATION). In all other cases, keep the default value, Default codec duration (DEFAULT_CODEC_DURATION). When you choose Match video duration, MediaConvert pads the output audio streams with silence or trims them to ensure that the total duration of each audio stream is at least as long as the total duration of the video stream. After padding or trimming, the audio stream duration is no more than one frame longer than the video stream. MediaConvert applies audio padding or trimming only to the end of the last segment of the output. For unsegmented outputs, MediaConvert adds padding only to the end of the file. When you keep the default value, any minor discrepancies between audio and video duration will depend on your output audio codec.

pub fn set_audio_duration(self, input: Option<M2tsAudioDuration>) -> Self

pub fn audio_frames_per_pes(self, input: i32) -> Self

The number of audio frames to insert for each PES packet.

pub fn set_audio_frames_per_pes(self, input: Option<i32>) -> Self

pub fn audio_pids(self, input: impl Into<i32>) -> Self

pub fn set_audio_pids(self, input: Option<Vec<i32>>) -> Self

pub fn bitrate(self, input: i32) -> Self

Specify the output bitrate of the transport stream in bits per second. Setting to 0 lets the muxer automatically determine the appropriate bitrate. Other common values are 3750000, 7500000, and 15000000.

pub fn set_bitrate(self, input: Option<i32>) -> Self

pub fn buffer_model(self, input: M2tsBufferModel) -> Self

Controls what buffer model to use for accurate interleaving. If set to MULTIPLEX, use multiplex buffer model. If set to NONE, this can lead to lower latency, but low-memory devices may not be able to play back the stream without interruptions.

pub fn set_buffer_model(self, input: Option<M2tsBufferModel>) -> Self

pub fn data_pts_control(self, input: M2tsDataPtsControl) -> Self

If you select ALIGN_TO_VIDEO, MediaConvert writes captions and data packets with Presentation Timestamp (PTS) values greater than or equal to the first video packet PTS (MediaConvert drops captions and data packets with lesser PTS values). Keep the default value (AUTO) to allow all PTS values.

pub fn set_data_pts_control(self, input: Option<M2tsDataPtsControl>) -> Self

pub fn dvb_nit_settings(self, input: DvbNitSettings) -> Self

Use these settings to insert a DVB Network Information Table (NIT) in the transport stream of this output. When you work directly in your JSON job specification, include this object only when your job has a transport stream output and the container settings contain the object M2tsSettings.

pub fn set_dvb_nit_settings(self, input: Option<DvbNitSettings>) -> Self

pub fn dvb_sdt_settings(self, input: DvbSdtSettings) -> Self

Use these settings to insert a DVB Service Description Table (SDT) in the transport stream of this output. When you work directly in your JSON job specification, include this object only when your job has a transport stream output and the container settings contain the object M2tsSettings.

pub fn set_dvb_sdt_settings(self, input: Option<DvbSdtSettings>) -> Self

pub fn dvb_sub_pids(self, input: impl Into<i32>) -> Self

pub fn set_dvb_sub_pids(self, input: Option<Vec<i32>>) -> Self

pub fn dvb_tdt_settings(self, input: DvbTdtSettings) -> Self

Use these settings to insert a DVB Time and Date Table (TDT) in the transport stream of this output. When you work directly in your JSON job specification, include this object only when your job has a transport stream output and the container settings contain the object M2tsSettings.

pub fn set_dvb_tdt_settings(self, input: Option<DvbTdtSettings>) -> Self

pub fn dvb_teletext_pid(self, input: i32) -> Self

Specify the packet identifier (PID) for DVB teletext data you include in this output. Default is 499.

pub fn set_dvb_teletext_pid(self, input: Option<i32>) -> Self

pub fn ebp_audio_interval(self, input: M2tsEbpAudioInterval) -> Self

When set to VIDEO_AND_FIXED_INTERVALS, audio EBP markers will be added to partitions 3 and 4. The interval between these additional markers will be fixed, and will be slightly shorter than the video EBP marker interval. When set to VIDEO_INTERVAL, these additional markers will not be inserted. Only applicable when EBP segmentation markers are is selected (segmentationMarkers is EBP or EBP_LEGACY).

pub fn set_ebp_audio_interval(self, input: Option<M2tsEbpAudioInterval>) -> Self

pub fn ebp_placement(self, input: M2tsEbpPlacement) -> Self

Selects which PIDs to place EBP markers on. They can either be placed only on the video PID, or on both the video PID and all audio PIDs. Only applicable when EBP segmentation markers are is selected (segmentationMarkers is EBP or EBP_LEGACY).

pub fn set_ebp_placement(self, input: Option<M2tsEbpPlacement>) -> Self

pub fn es_rate_in_pes(self, input: M2tsEsRateInPes) -> Self

Controls whether to include the ES Rate field in the PES header.

pub fn set_es_rate_in_pes(self, input: Option<M2tsEsRateInPes>) -> Self

pub fn force_ts_video_ebp_order(self, input: M2tsForceTsVideoEbpOrder) -> Self

Keep the default value (DEFAULT) unless you know that your audio EBP markers are incorrectly appearing before your video EBP markers. To correct this problem, set this value to Force (FORCE).

pub fn set_force_ts_video_ebp_order(
self,
input: Option<M2tsForceTsVideoEbpOrder>
) -> Self

pub fn fragment_time(self, input: f64) -> Self

The length, in seconds, of each fragment. Only used with EBP markers.

pub fn set_fragment_time(self, input: Option<f64>) -> Self

pub fn max_pcr_interval(self, input: i32) -> Self

Specify the maximum time, in milliseconds, between Program Clock References (PCRs) inserted into the transport stream.

pub fn set_max_pcr_interval(self, input: Option<i32>) -> Self

pub fn min_ebp_interval(self, input: i32) -> Self

When set, enforces that Encoder Boundary Points do not come within the specified time interval of each other by looking ahead at input video. If another EBP is going to come in within the specified time interval, the current EBP is not emitted, and the segment is “stretched” to the next marker. The lookahead value does not add latency to the system. The Live Event must be configured elsewhere to create sufficient latency to make the lookahead accurate.

pub fn set_min_ebp_interval(self, input: Option<i32>) -> Self

pub fn nielsen_id3(self, input: M2tsNielsenId3) -> Self

If INSERT, Nielsen inaudible tones for media tracking will be detected in the input audio and an equivalent ID3 tag will be inserted in the output.

pub fn set_nielsen_id3(self, input: Option<M2tsNielsenId3>) -> Self

pub fn null_packet_bitrate(self, input: f64) -> Self

Value in bits per second of extra null packets to insert into the transport stream. This can be used if a downstream encryption system requires periodic null packets.

pub fn set_null_packet_bitrate(self, input: Option<f64>) -> Self

pub fn pat_interval(self, input: i32) -> Self

The number of milliseconds between instances of this table in the output transport stream.

pub fn set_pat_interval(self, input: Option<i32>) -> Self

pub fn pcr_control(self, input: M2tsPcrControl) -> Self

When set to PCR_EVERY_PES_PACKET, a Program Clock Reference value is inserted for every Packetized Elementary Stream (PES) header. This is effective only when the PCR PID is the same as the video or audio elementary stream.

pub fn set_pcr_control(self, input: Option<M2tsPcrControl>) -> Self

pub fn pcr_pid(self, input: i32) -> Self

Specify the packet identifier (PID) for the program clock reference (PCR) in this output. If you do not specify a value, the service will use the value for Video PID (VideoPid).

pub fn set_pcr_pid(self, input: Option<i32>) -> Self

pub fn pmt_interval(self, input: i32) -> Self

Specify the number of milliseconds between instances of the program map table (PMT) in the output transport stream.

pub fn set_pmt_interval(self, input: Option<i32>) -> Self

pub fn pmt_pid(self, input: i32) -> Self

Specify the packet identifier (PID) for the program map table (PMT) itself. Default is 480.

pub fn set_pmt_pid(self, input: Option<i32>) -> Self

pub fn private_metadata_pid(self, input: i32) -> Self

Specify the packet identifier (PID) of the private metadata stream. Default is 503.

pub fn set_private_metadata_pid(self, input: Option<i32>) -> Self

pub fn program_number(self, input: i32) -> Self

Use Program number (programNumber) to specify the program number used in the program map table (PMT) for this output. Default is 1. Program numbers and program map tables are parts of MPEG-2 transport stream containers, used for organizing data.

pub fn set_program_number(self, input: Option<i32>) -> Self

pub fn rate_mode(self, input: M2tsRateMode) -> Self

When set to CBR, inserts null packets into transport stream to fill specified bitrate. When set to VBR, the bitrate setting acts as the maximum bitrate, but the output will not be padded up to that bitrate.

pub fn set_rate_mode(self, input: Option<M2tsRateMode>) -> Self

pub fn scte35_esam(self, input: M2tsScte35Esam) -> Self

Include this in your job settings to put SCTE-35 markers in your HLS and transport stream outputs at the insertion points that you specify in an ESAM XML document. Provide the document in the setting SCC XML (sccXml).

pub fn set_scte35_esam(self, input: Option<M2tsScte35Esam>) -> Self

pub fn scte35_pid(self, input: i32) -> Self

Specify the packet identifier (PID) of the SCTE-35 stream in the transport stream.

pub fn set_scte35_pid(self, input: Option<i32>) -> Self

pub fn scte35_source(self, input: M2tsScte35Source) -> Self

For SCTE-35 markers from your input– Choose Passthrough (PASSTHROUGH) if you want SCTE-35 markers that appear in your input to also appear in this output. Choose None (NONE) if you don’t want SCTE-35 markers in this output. For SCTE-35 markers from an ESAM XML document– Choose None (NONE). Also provide the ESAM XML as a string in the setting Signal processing notification XML (sccXml). Also enable ESAM SCTE-35 (include the property scte35Esam).

pub fn set_scte35_source(self, input: Option<M2tsScte35Source>) -> Self

pub fn segmentation_markers(self, input: M2tsSegmentationMarkers) -> Self

Inserts segmentation markers at each segmentation_time period. rai_segstart sets the Random Access Indicator bit in the adaptation field. rai_adapt sets the RAI bit and adds the current timecode in the private data bytes. psi_segstart inserts PAT and PMT tables at the start of segments. ebp adds Encoder Boundary Point information to the adaptation field as per OpenCable specification OC-SP-EBP-I01-130118. ebp_legacy adds Encoder Boundary Point information to the adaptation field using a legacy proprietary format.

pub fn set_segmentation_markers(
self,
input: Option<M2tsSegmentationMarkers>
) -> Self

pub fn segmentation_style(self, input: M2tsSegmentationStyle) -> Self

The segmentation style parameter controls how segmentation markers are inserted into the transport stream. With avails, it is possible that segments may be truncated, which can influence where future segmentation markers are inserted. When a segmentation style of “reset_cadence” is selected and a segment is truncated due to an avail, we will reset the segmentation cadence. This means the subsequent segment will have a duration of of $segmentation_time seconds. When a segmentation style of “maintain_cadence” is selected and a segment is truncated due to an avail, we will not reset the segmentation cadence. This means the subsequent segment will likely be truncated as well. However, all segments after that will have a duration of $segmentation_time seconds. Note that EBP lookahead is a slight exception to this rule.

pub fn set_segmentation_style(
self,
input: Option<M2tsSegmentationStyle>
) -> Self

pub fn segmentation_time(self, input: f64) -> Self

Specify the length, in seconds, of each segment. Required unless markers is set to none.

pub fn set_segmentation_time(self, input: Option<f64>) -> Self

pub fn timed_metadata_pid(self, input: i32) -> Self

Specify the packet identifier (PID) for timed metadata in this output. Default is 502.

pub fn set_timed_metadata_pid(self, input: Option<i32>) -> Self

pub fn transport_stream_id(self, input: i32) -> Self

Specify the ID for the transport stream itself in the program map table for this output. Transport stream IDs and program map tables are parts of MPEG-2 transport stream containers, used for organizing data.

pub fn set_transport_stream_id(self, input: Option<i32>) -> Self

pub fn video_pid(self, input: i32) -> Self

Specify the packet identifier (PID) of the elementary video stream in the transport stream.

pub fn set_video_pid(self, input: Option<i32>) -> Self

pub fn build(self) -> M2tsSettings

Consumes the builder and constructs a M2tsSettings