Struct aws_sdk_mediaconvert::model::ms_smooth_group_settings::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for MsSmoothGroupSettings

Implementations

impl Builder

pub fn additional_manifests(
self,
input: impl Into<MsSmoothAdditionalManifest>
) -> Self

pub fn set_additional_manifests(
self,
input: Option<Vec<MsSmoothAdditionalManifest>>
) -> Self

pub fn audio_deduplication(self, input: MsSmoothAudioDeduplication) -> Self

COMBINE_DUPLICATE_STREAMS combines identical audio encoding settings across a Microsoft Smooth output group into a single audio stream.

pub fn set_audio_deduplication(
self,
input: Option<MsSmoothAudioDeduplication>
) -> Self

pub fn destination(self, input: impl Into<String>) -> Self

Use Destination (Destination) to specify the S3 output location and the output filename base. Destination accepts format identifiers. If you do not specify the base filename in the URI, the service will use the filename of the input file. If your job has multiple inputs, the service uses the filename of the first input file.

pub fn set_destination(self, input: Option<String>) -> Self

pub fn destination_settings(self, input: DestinationSettings) -> Self

Settings associated with the destination. Will vary based on the type of destination

pub fn set_destination_settings(
self,
input: Option<DestinationSettings>
) -> Self

pub fn encryption(self, input: MsSmoothEncryptionSettings) -> Self

If you are using DRM, set DRM System (MsSmoothEncryptionSettings) to specify the value SpekeKeyProvider.

pub fn set_encryption(self, input: Option<MsSmoothEncryptionSettings>) -> Self

pub fn fragment_length(self, input: i32) -> Self

Specify how you want MediaConvert to determine the fragment length. Choose Exact (EXACT) to have the encoder use the exact length that you specify with the setting Fragment length (FragmentLength). This might result in extra I-frames. Choose Multiple of GOP (GOP_MULTIPLE) to have the encoder round up the segment lengths to match the next GOP boundary.

pub fn set_fragment_length(self, input: Option<i32>) -> Self

pub fn fragment_length_control(
self,
input: MsSmoothFragmentLengthControl
) -> Self

Specify how you want MediaConvert to determine the fragment length. Choose Exact (EXACT) to have the encoder use the exact length that you specify with the setting Fragment length (FragmentLength). This might result in extra I-frames. Choose Multiple of GOP (GOP_MULTIPLE) to have the encoder round up the segment lengths to match the next GOP boundary.

pub fn set_fragment_length_control(
self,
input: Option<MsSmoothFragmentLengthControl>
) -> Self

pub fn manifest_encoding(self, input: MsSmoothManifestEncoding) -> Self

Use Manifest encoding (MsSmoothManifestEncoding) to specify the encoding format for the server and client manifest. Valid options are utf8 and utf16.

pub fn set_manifest_encoding(
self,
input: Option<MsSmoothManifestEncoding>
) -> Self

pub fn build(self) -> MsSmoothGroupSettings

Consumes the builder and constructs a MsSmoothGroupSettings