Struct aws_sdk_mediaconvert::model::CmafGroupSettings

#[non_exhaustive]
pub struct CmafGroupSettings {

    pub additional_manifests: Option<Vec<CmafAdditionalManifest>>,
    pub base_url: Option<String>,
    pub client_cache: Option<CmafClientCache>,
    pub codec_specification: Option<CmafCodecSpecification>,
    pub destination: Option<String>,
    pub destination_settings: Option<DestinationSettings>,
    pub encryption: Option<CmafEncryptionSettings>,
    pub fragment_length: i32,
    pub image_based_trick_play: Option<CmafImageBasedTrickPlay>,
    pub manifest_compression: Option<CmafManifestCompression>,
    pub manifest_duration_format: Option<CmafManifestDurationFormat>,
    pub min_buffer_time: i32,
    pub min_final_segment_length: f64,
    pub mpd_profile: Option<CmafMpdProfile>,
    pub pts_offset_handling_for_b_frames: Option<CmafPtsOffsetHandlingForBFrames>,
    pub segment_control: Option<CmafSegmentControl>,
    pub segment_length: i32,
    pub segment_length_control: Option<CmafSegmentLengthControl>,
    pub stream_inf_resolution: Option<CmafStreamInfResolution>,
    pub target_duration_compatibility_mode: Option<CmafTargetDurationCompatibilityMode>,
    pub write_dash_manifest: Option<CmafWriteDashManifest>,
    pub write_hls_manifest: Option<CmafWriteHlsManifest>,
    pub write_segment_timeline_in_representation: Option<CmafWriteSegmentTimelineInRepresentation>,

}

Settings related to your CMAF output package. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/outputs-file-ABR.html. When you work directly in your JSON job specification, include this object and any required children when you set Type, under OutputGroupSettings, to CMAF_GROUP_SETTINGS.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

additional_manifests: Option<Vec<CmafAdditionalManifest>>

By default, the service creates one top-level .m3u8 HLS manifest and one top -level .mpd DASH manifest for each CMAF output group in your job. These default manifests reference every output in the output group. To create additional top-level manifests that reference a subset of the outputs in the output group, specify a list of them here. For each additional manifest that you specify, the service creates one HLS manifest and one DASH manifest.

base_url: Option<String>

A partial URI prefix that will be put in the manifest file at the top level BaseURL element. Can be used if streams are delivered from a different URL than the manifest file.

client_cache: Option<CmafClientCache>

Disable this setting only when your workflow requires the #EXT-X-ALLOW-CACHE:no tag. Otherwise, keep the default value Enabled (ENABLED) and control caching in your video distribution set up. For example, use the Cache-Control http header.

codec_specification: Option<CmafCodecSpecification>

Specification to use (RFC-6381 or the default RFC-4281) during m3u8 playlist generation.

destination: Option<String>

Use Destination (Destination) to specify the S3 output location and the output filename base. Destination accepts format identifiers. If you do not specify the base filename in the URI, the service will use the filename of the input file. If your job has multiple inputs, the service uses the filename of the first input file.

destination_settings: Option<DestinationSettings>

Settings associated with the destination. Will vary based on the type of destination

encryption: Option<CmafEncryptionSettings>

DRM settings.

fragment_length: i32

Specify the length, in whole seconds, of the mp4 fragments. When you don’t specify a value, MediaConvert defaults to 2. Related setting: Use Fragment length control (FragmentLengthControl) to specify whether the encoder enforces this value strictly.

image_based_trick_play: Option<CmafImageBasedTrickPlay>

Specify whether MediaConvert generates images for trick play. Keep the default value, None (NONE), to not generate any images. Choose Thumbnail (THUMBNAIL) to generate tiled thumbnails. Choose Thumbnail and full frame (THUMBNAIL_AND_FULLFRAME) to generate tiled thumbnails and full-resolution images of single frames. When you enable Write HLS manifest (WriteHlsManifest), MediaConvert creates a child manifest for each set of images that you generate and adds corresponding entries to the parent manifest. When you enable Write DASH manifest (WriteDashManifest), MediaConvert adds an entry in the .mpd manifest for each set of images that you generate. A common application for these images is Roku trick mode. The thumbnails and full-frame images that MediaConvert creates with this feature are compatible with this Roku specification: https://developer.roku.com/docs/developer-program/media-playback/trick-mode/hls-and-dash.md

manifest_compression: Option<CmafManifestCompression>

When set to GZIP, compresses HLS playlist.

manifest_duration_format: Option<CmafManifestDurationFormat>

Indicates whether the output manifest should use floating point values for segment duration.

min_buffer_time: i32

Minimum time of initially buffered media that is needed to ensure smooth playout.

min_final_segment_length: f64

Keep this setting at the default value of 0, unless you are troubleshooting a problem with how devices play back the end of your video asset. If you know that player devices are hanging on the final segment of your video because the length of your final segment is too short, use this setting to specify a minimum final segment length, in seconds. Choose a value that is greater than or equal to 1 and less than your segment length. When you specify a value for this setting, the encoder will combine any final segment that is shorter than the length that you specify with the previous segment. For example, your segment length is 3 seconds and your final segment is .5 seconds without a minimum final segment length; when you set the minimum final segment length to 1, your final segment is 3.5 seconds.

mpd_profile: Option<CmafMpdProfile>

Specify whether your DASH profile is on-demand or main. When you choose Main profile (MAIN_PROFILE), the service signals urn:mpeg:dash:profile:isoff-main:2011 in your .mpd DASH manifest. When you choose On-demand (ON_DEMAND_PROFILE), the service signals urn:mpeg:dash:profile:isoff-on-demand:2011 in your .mpd. When you choose On-demand, you must also set the output group setting Segment control (SegmentControl) to Single file (SINGLE_FILE).

pts_offset_handling_for_b_frames: Option<CmafPtsOffsetHandlingForBFrames>

Use this setting only when your output video stream has B-frames, which causes the initial presentation time stamp (PTS) to be offset from the initial decode time stamp (DTS). Specify how MediaConvert handles PTS when writing time stamps in output DASH manifests. Choose Match initial PTS (MATCH_INITIAL_PTS) when you want MediaConvert to use the initial PTS as the first time stamp in the manifest. Choose Zero-based (ZERO_BASED) to have MediaConvert ignore the initial PTS in the video stream and instead write the initial time stamp as zero in the manifest. For outputs that don’t have B-frames, the time stamps in your DASH manifests start at zero regardless of your choice here.

segment_control: Option<CmafSegmentControl>

When set to SINGLE_FILE, a single output file is generated, which is internally segmented using the Fragment Length and Segment Length. When set to SEGMENTED_FILES, separate segment files will be created.

segment_length: i32

Specify the length, in whole seconds, of each segment. When you don’t specify a value, MediaConvert defaults to 10. Related settings: Use Segment length control (SegmentLengthControl) to specify whether the encoder enforces this value strictly. Use Segment control (CmafSegmentControl) to specify whether MediaConvert creates separate segment files or one content file that has metadata to mark the segment boundaries.

segment_length_control: Option<CmafSegmentLengthControl>

Specify how you want MediaConvert to determine the segment length. Choose Exact (EXACT) to have the encoder use the exact length that you specify with the setting Segment length (SegmentLength). This might result in extra I-frames. Choose Multiple of GOP (GOP_MULTIPLE) to have the encoder round up the segment lengths to match the next GOP boundary.

stream_inf_resolution: Option<CmafStreamInfResolution>

Include or exclude RESOLUTION attribute for video in EXT-X-STREAM-INF tag of variant manifest.

target_duration_compatibility_mode: Option<CmafTargetDurationCompatibilityMode>

When set to LEGACY, the segment target duration is always rounded up to the nearest integer value above its current value in seconds. When set to SPEC\_COMPLIANT, the segment target duration is rounded up to the nearest integer value if fraction seconds are greater than or equal to 0.5 (>= 0.5) and rounded down if less than 0.5 (< 0.5). You may need to use LEGACY if your client needs to ensure that the target duration is always longer than the actual duration of the segment. Some older players may experience interrupted playback when the actual duration of a track in a segment is longer than the target duration.

write_dash_manifest: Option<CmafWriteDashManifest>

When set to ENABLED, a DASH MPD manifest will be generated for this output.

write_hls_manifest: Option<CmafWriteHlsManifest>

When set to ENABLED, an Apple HLS manifest will be generated for this output.

write_segment_timeline_in_representation: Option<CmafWriteSegmentTimelineInRepresentation>

When you enable Precise segment duration in DASH manifests (writeSegmentTimelineInRepresentation), your DASH manifest shows precise segment durations. The segment duration information appears inside the SegmentTimeline element, inside SegmentTemplate at the Representation level. When this feature isn’t enabled, the segment durations in your DASH manifest are approximate. The segment duration information appears in the duration attribute of the SegmentTemplate element.

Implementations

impl CmafGroupSettings

pub fn builder() -> Builder

Creates a new builder-style object to manufacture CmafGroupSettings