Struct aws_sdk_mediaconvert::model::DashIsoGroupSettings

#[non_exhaustive]
pub struct DashIsoGroupSettings {

    pub additional_manifests: Option<Vec<DashAdditionalManifest>>,
    pub audio_channel_config_scheme_id_uri: Option<DashIsoGroupAudioChannelConfigSchemeIdUri>,
    pub base_url: Option<String>,
    pub destination: Option<String>,
    pub destination_settings: Option<DestinationSettings>,
    pub encryption: Option<DashIsoEncryptionSettings>,
    pub fragment_length: i32,
    pub hbbtv_compliance: Option<DashIsoHbbtvCompliance>,
    pub image_based_trick_play: Option<DashIsoImageBasedTrickPlay>,
    pub min_buffer_time: i32,
    pub min_final_segment_length: f64,
    pub mpd_profile: Option<DashIsoMpdProfile>,
    pub pts_offset_handling_for_b_frames: Option<DashIsoPtsOffsetHandlingForBFrames>,
    pub segment_control: Option<DashIsoSegmentControl>,
    pub segment_length: i32,
    pub segment_length_control: Option<DashIsoSegmentLengthControl>,
    pub write_segment_timeline_in_representation: Option<DashIsoWriteSegmentTimelineInRepresentation>,

}

Settings related to your DASH output package. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/outputs-file-ABR.html. When you work directly in your JSON job specification, include this object and any required children when you set Type, under OutputGroupSettings, to DASH_ISO_GROUP_SETTINGS.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

additional_manifests: Option<Vec<DashAdditionalManifest>>

By default, the service creates one .mpd DASH manifest for each DASH ISO output group in your job. This default manifest references every output in the output group. To create additional DASH manifests that reference a subset of the outputs in the output group, specify a list of them here.

audio_channel_config_scheme_id_uri: Option<DashIsoGroupAudioChannelConfigSchemeIdUri>

Use this setting only when your audio codec is a Dolby one (AC3, EAC3, or Atmos) and your downstream workflow requires that your DASH manifest use the Dolby channel configuration tag, rather than the MPEG one. For example, you might need to use this to make dynamic ad insertion work. Specify which audio channel configuration scheme ID URI MediaConvert writes in your DASH manifest. Keep the default value, MPEG channel configuration (MPEG_CHANNEL_CONFIGURATION), to have MediaConvert write this: urn:mpeg:mpegB:cicp:ChannelConfiguration. Choose Dolby channel configuration (DOLBY_CHANNEL_CONFIGURATION) to have MediaConvert write this instead: tag:dolby.com,2014:dash:audio_channel_configuration:2011.

base_url: Option<String>

A partial URI prefix that will be put in the manifest (.mpd) file at the top level BaseURL element. Can be used if streams are delivered from a different URL than the manifest file.

destination: Option<String>

Use Destination (Destination) to specify the S3 output location and the output filename base. Destination accepts format identifiers. If you do not specify the base filename in the URI, the service will use the filename of the input file. If your job has multiple inputs, the service uses the filename of the first input file.

destination_settings: Option<DestinationSettings>

Settings associated with the destination. Will vary based on the type of destination

encryption: Option<DashIsoEncryptionSettings>

DRM settings.

fragment_length: i32

Length of fragments to generate (in seconds). Fragment length must be compatible with GOP size and Framerate. Note that fragments will end on the next keyframe after this number of seconds, so actual fragment length may be longer. When Emit Single File is checked, the fragmentation is internal to a single output file and it does not cause the creation of many output files as in other output types.

hbbtv_compliance: Option<DashIsoHbbtvCompliance>

Supports HbbTV specification as indicated

image_based_trick_play: Option<DashIsoImageBasedTrickPlay>

Specify whether MediaConvert generates images for trick play. Keep the default value, None (NONE), to not generate any images. Choose Thumbnail (THUMBNAIL) to generate tiled thumbnails. Choose Thumbnail and full frame (THUMBNAIL_AND_FULLFRAME) to generate tiled thumbnails and full-resolution images of single frames. MediaConvert adds an entry in the .mpd manifest for each set of images that you generate. A common application for these images is Roku trick mode. The thumbnails and full-frame images that MediaConvert creates with this feature are compatible with this Roku specification: https://developer.roku.com/docs/developer-program/media-playback/trick-mode/hls-and-dash.md

min_buffer_time: i32

Minimum time of initially buffered media that is needed to ensure smooth playout.

min_final_segment_length: f64

Keep this setting at the default value of 0, unless you are troubleshooting a problem with how devices play back the end of your video asset. If you know that player devices are hanging on the final segment of your video because the length of your final segment is too short, use this setting to specify a minimum final segment length, in seconds. Choose a value that is greater than or equal to 1 and less than your segment length. When you specify a value for this setting, the encoder will combine any final segment that is shorter than the length that you specify with the previous segment. For example, your segment length is 3 seconds and your final segment is .5 seconds without a minimum final segment length; when you set the minimum final segment length to 1, your final segment is 3.5 seconds.

mpd_profile: Option<DashIsoMpdProfile>

Specify whether your DASH profile is on-demand or main. When you choose Main profile (MAIN_PROFILE), the service signals urn:mpeg:dash:profile:isoff-main:2011 in your .mpd DASH manifest. When you choose On-demand (ON_DEMAND_PROFILE), the service signals urn:mpeg:dash:profile:isoff-on-demand:2011 in your .mpd. When you choose On-demand, you must also set the output group setting Segment control (SegmentControl) to Single file (SINGLE_FILE).

pts_offset_handling_for_b_frames: Option<DashIsoPtsOffsetHandlingForBFrames>

Use this setting only when your output video stream has B-frames, which causes the initial presentation time stamp (PTS) to be offset from the initial decode time stamp (DTS). Specify how MediaConvert handles PTS when writing time stamps in output DASH manifests. Choose Match initial PTS (MATCH_INITIAL_PTS) when you want MediaConvert to use the initial PTS as the first time stamp in the manifest. Choose Zero-based (ZERO_BASED) to have MediaConvert ignore the initial PTS in the video stream and instead write the initial time stamp as zero in the manifest. For outputs that don’t have B-frames, the time stamps in your DASH manifests start at zero regardless of your choice here.

segment_control: Option<DashIsoSegmentControl>

When set to SINGLE_FILE, a single output file is generated, which is internally segmented using the Fragment Length and Segment Length. When set to SEGMENTED_FILES, separate segment files will be created.

segment_length: i32

Specify the length, in whole seconds, of each segment. When you don’t specify a value, MediaConvert defaults to 30. Related settings: Use Segment length control (SegmentLengthControl) to specify whether the encoder enforces this value strictly. Use Segment control (DashIsoSegmentControl) to specify whether MediaConvert creates separate segment files or one content file that has metadata to mark the segment boundaries.

segment_length_control: Option<DashIsoSegmentLengthControl>

Specify how you want MediaConvert to determine the segment length. Choose Exact (EXACT) to have the encoder use the exact length that you specify with the setting Segment length (SegmentLength). This might result in extra I-frames. Choose Multiple of GOP (GOP_MULTIPLE) to have the encoder round up the segment lengths to match the next GOP boundary.

write_segment_timeline_in_representation: Option<DashIsoWriteSegmentTimelineInRepresentation>

If you get an HTTP error in the 400 range when you play back your DASH output, enable this setting and run your transcoding job again. When you enable this setting, the service writes precise segment durations in the DASH manifest. The segment duration information appears inside the SegmentTimeline element, inside SegmentTemplate at the Representation level. When you don’t enable this setting, the service writes approximate segment durations in your DASH manifest.

Implementations

impl DashIsoGroupSettings

pub fn builder() -> Builder

Creates a new builder-style object to manufacture DashIsoGroupSettings