Struct aws_sdk_mediaconvert::model::Eac3AtmosSettings

#[non_exhaustive]
pub struct Eac3AtmosSettings {

    pub bitrate: i32,
    pub bitstream_mode: Option<Eac3AtmosBitstreamMode>,
    pub coding_mode: Option<Eac3AtmosCodingMode>,
    pub dialogue_intelligence: Option<Eac3AtmosDialogueIntelligence>,
    pub downmix_control: Option<Eac3AtmosDownmixControl>,
    pub dynamic_range_compression_line: Option<Eac3AtmosDynamicRangeCompressionLine>,
    pub dynamic_range_compression_rf: Option<Eac3AtmosDynamicRangeCompressionRf>,
    pub dynamic_range_control: Option<Eac3AtmosDynamicRangeControl>,
    pub lo_ro_center_mix_level: f64,
    pub lo_ro_surround_mix_level: f64,
    pub lt_rt_center_mix_level: f64,
    pub lt_rt_surround_mix_level: f64,
    pub metering_mode: Option<Eac3AtmosMeteringMode>,
    pub sample_rate: i32,
    pub speech_threshold: i32,
    pub stereo_downmix: Option<Eac3AtmosStereoDownmix>,
    pub surround_ex_mode: Option<Eac3AtmosSurroundExMode>,

}

Required when you set (Codec) under (AudioDescriptions)>(CodecSettings) to the value EAC3_ATMOS.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

bitrate: i32

Specify the average bitrate for this output in bits per second. Valid values: 384k, 448k, 576k, 640k, 768k, 1024k Default value: 448k Note that MediaConvert supports 384k only with channel-based immersive (CBI) 7.1.4 and 5.1.4 inputs. For CBI 9.1.6 and other input types, MediaConvert automatically increases your output bitrate to 448k.

bitstream_mode: Option<Eac3AtmosBitstreamMode>

Specify the bitstream mode for the E-AC-3 stream that the encoder emits. For more information about the EAC3 bitstream mode, see ATSC A/52-2012 (Annex E).

coding_mode: Option<Eac3AtmosCodingMode>

The coding mode for Dolby Digital Plus JOC (Atmos).

dialogue_intelligence: Option<Eac3AtmosDialogueIntelligence>

Enable Dolby Dialogue Intelligence to adjust loudness based on dialogue analysis.

downmix_control: Option<Eac3AtmosDownmixControl>

Specify whether MediaConvert should use any downmix metadata from your input file. Keep the default value, Custom (SPECIFIED) to provide downmix values in your job settings. Choose Follow source (INITIALIZE_FROM_SOURCE) to use the metadata from your input. Related settings–Use these settings to specify your downmix values: Left only/Right only surround (LoRoSurroundMixLevel), Left total/Right total surround (LtRtSurroundMixLevel), Left total/Right total center (LtRtCenterMixLevel), Left only/Right only center (LoRoCenterMixLevel), and Stereo downmix (StereoDownmix). When you keep Custom (SPECIFIED) for Downmix control (DownmixControl) and you don’t specify values for the related settings, MediaConvert uses default values for those settings.

dynamic_range_compression_line: Option<Eac3AtmosDynamicRangeCompressionLine>

Choose the Dolby dynamic range control (DRC) profile that MediaConvert uses when encoding the metadata in the Dolby stream for the line operating mode. Default value: Film light (ATMOS_STORAGE_DDP_COMPR_FILM_LIGHT) Related setting: To have MediaConvert use the value you specify here, keep the default value, Custom (SPECIFIED) for the setting Dynamic range control (DynamicRangeControl). Otherwise, MediaConvert ignores Dynamic range compression line (DynamicRangeCompressionLine). For information about the Dolby DRC operating modes and profiles, see the Dynamic Range Control chapter of the Dolby Metadata Guide at https://developer.dolby.com/globalassets/professional/documents/dolby-metadata-guide.pdf.

dynamic_range_compression_rf: Option<Eac3AtmosDynamicRangeCompressionRf>

Choose the Dolby dynamic range control (DRC) profile that MediaConvert uses when encoding the metadata in the Dolby stream for the RF operating mode. Default value: Film light (ATMOS_STORAGE_DDP_COMPR_FILM_LIGHT) Related setting: To have MediaConvert use the value you specify here, keep the default value, Custom (SPECIFIED) for the setting Dynamic range control (DynamicRangeControl). Otherwise, MediaConvert ignores Dynamic range compression RF (DynamicRangeCompressionRf). For information about the Dolby DRC operating modes and profiles, see the Dynamic Range Control chapter of the Dolby Metadata Guide at https://developer.dolby.com/globalassets/professional/documents/dolby-metadata-guide.pdf.

dynamic_range_control: Option<Eac3AtmosDynamicRangeControl>

Specify whether MediaConvert should use any dynamic range control metadata from your input file. Keep the default value, Custom (SPECIFIED), to provide dynamic range control values in your job settings. Choose Follow source (INITIALIZE_FROM_SOURCE) to use the metadata from your input. Related settings–Use these settings to specify your dynamic range control values: Dynamic range compression line (DynamicRangeCompressionLine) and Dynamic range compression RF (DynamicRangeCompressionRf). When you keep the value Custom (SPECIFIED) for Dynamic range control (DynamicRangeControl) and you don’t specify values for the related settings, MediaConvert uses default values for those settings.

lo_ro_center_mix_level: f64

Specify a value for the following Dolby Atmos setting: Left only/Right only center mix (Lo/Ro center). MediaConvert uses this value for downmixing. Default value: -3 dB (ATMOS_STORAGE_DDP_MIXLEV_MINUS_3_DB). Valid values: 3.0, 1.5, 0.0, -1.5, -3.0, -4.5, and -6.0. Related setting: How the service uses this value depends on the value that you choose for Stereo downmix (Eac3AtmosStereoDownmix). Related setting: To have MediaConvert use this value, keep the default value, Custom (SPECIFIED) for the setting Downmix control (DownmixControl). Otherwise, MediaConvert ignores Left only/Right only center (LoRoCenterMixLevel).

lo_ro_surround_mix_level: f64

Specify a value for the following Dolby Atmos setting: Left only/Right only (Lo/Ro surround). MediaConvert uses this value for downmixing. Default value: -3 dB (ATMOS_STORAGE_DDP_MIXLEV_MINUS_3_DB). Valid values: -1.5, -3.0, -4.5, -6.0, and -60. The value -60 mutes the channel. Related setting: How the service uses this value depends on the value that you choose for Stereo downmix (Eac3AtmosStereoDownmix). Related setting: To have MediaConvert use this value, keep the default value, Custom (SPECIFIED) for the setting Downmix control (DownmixControl). Otherwise, MediaConvert ignores Left only/Right only surround (LoRoSurroundMixLevel).

lt_rt_center_mix_level: f64

Specify a value for the following Dolby Atmos setting: Left total/Right total center mix (Lt/Rt center). MediaConvert uses this value for downmixing. Default value: -3 dB (ATMOS_STORAGE_DDP_MIXLEV_MINUS_3_DB) Valid values: 3.0, 1.5, 0.0, -1.5, -3.0, -4.5, and -6.0. Related setting: How the service uses this value depends on the value that you choose for Stereo downmix (Eac3AtmosStereoDownmix). Related setting: To have MediaConvert use this value, keep the default value, Custom (SPECIFIED) for the setting Downmix control (DownmixControl). Otherwise, MediaConvert ignores Left total/Right total center (LtRtCenterMixLevel).

lt_rt_surround_mix_level: f64

Specify a value for the following Dolby Atmos setting: Left total/Right total surround mix (Lt/Rt surround). MediaConvert uses this value for downmixing. Default value: -3 dB (ATMOS_STORAGE_DDP_MIXLEV_MINUS_3_DB) Valid values: -1.5, -3.0, -4.5, -6.0, and -60. The value -60 mutes the channel. Related setting: How the service uses this value depends on the value that you choose for Stereo downmix (Eac3AtmosStereoDownmix). Related setting: To have MediaConvert use this value, keep the default value, Custom (SPECIFIED) for the setting Downmix control (DownmixControl). Otherwise, the service ignores Left total/Right total surround (LtRtSurroundMixLevel).

metering_mode: Option<Eac3AtmosMeteringMode>

Choose how the service meters the loudness of your audio.

sample_rate: i32

This value is always 48000. It represents the sample rate in Hz.

speech_threshold: i32

Specify the percentage of audio content, from 0% to 100%, that must be speech in order for the encoder to use the measured speech loudness as the overall program loudness. Default value: 15%

stereo_downmix: Option<Eac3AtmosStereoDownmix>

Choose how the service does stereo downmixing. Default value: Not indicated (ATMOS_STORAGE_DDP_DMIXMOD_NOT_INDICATED) Related setting: To have MediaConvert use this value, keep the default value, Custom (SPECIFIED) for the setting Downmix control (DownmixControl). Otherwise, MediaConvert ignores Stereo downmix (StereoDownmix).

surround_ex_mode: Option<Eac3AtmosSurroundExMode>

Specify whether your input audio has an additional center rear surround channel matrix encoded into your left and right surround channels.

Implementations

impl Eac3AtmosSettings

pub fn builder() -> Builder

Creates a new builder-style object to manufacture Eac3AtmosSettings