Struct aws_sdk_mediaconvert::model::HlsGroupSettings

#[non_exhaustive]
pub struct HlsGroupSettings {

    pub ad_markers: Option<Vec<HlsAdMarkers>>,
    pub additional_manifests: Option<Vec<HlsAdditionalManifest>>,
    pub audio_only_header: Option<HlsAudioOnlyHeader>,
    pub base_url: Option<String>,
    pub caption_language_mappings: Option<Vec<HlsCaptionLanguageMapping>>,
    pub caption_language_setting: Option<HlsCaptionLanguageSetting>,
    pub client_cache: Option<HlsClientCache>,
    pub codec_specification: Option<HlsCodecSpecification>,
    pub destination: Option<String>,
    pub destination_settings: Option<DestinationSettings>,
    pub directory_structure: Option<HlsDirectoryStructure>,
    pub encryption: Option<HlsEncryptionSettings>,
    pub image_based_trick_play: Option<HlsImageBasedTrickPlay>,
    pub manifest_compression: Option<HlsManifestCompression>,
    pub manifest_duration_format: Option<HlsManifestDurationFormat>,
    pub min_final_segment_length: f64,
    pub min_segment_length: i32,
    pub output_selection: Option<HlsOutputSelection>,
    pub program_date_time: Option<HlsProgramDateTime>,
    pub program_date_time_period: i32,
    pub segment_control: Option<HlsSegmentControl>,
    pub segment_length: i32,
    pub segment_length_control: Option<HlsSegmentLengthControl>,
    pub segments_per_subdirectory: i32,
    pub stream_inf_resolution: Option<HlsStreamInfResolution>,
    pub target_duration_compatibility_mode: Option<HlsTargetDurationCompatibilityMode>,
    pub timed_metadata_id3_frame: Option<HlsTimedMetadataId3Frame>,
    pub timed_metadata_id3_period: i32,
    pub timestamp_delta_milliseconds: i32,

}

Settings related to your HLS output package. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/outputs-file-ABR.html. When you work directly in your JSON job specification, include this object and any required children when you set Type, under OutputGroupSettings, to HLS_GROUP_SETTINGS.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

ad_markers: Option<Vec<HlsAdMarkers>>

Choose one or more ad marker types to decorate your Apple HLS manifest. This setting does not determine whether SCTE-35 markers appear in the outputs themselves.

additional_manifests: Option<Vec<HlsAdditionalManifest>>

By default, the service creates one top-level .m3u8 HLS manifest for each HLS output group in your job. This default manifest references every output in the output group. To create additional top-level manifests that reference a subset of the outputs in the output group, specify a list of them here.

audio_only_header: Option<HlsAudioOnlyHeader>

Ignore this setting unless you are using FairPlay DRM with Verimatrix and you encounter playback issues. Keep the default value, Include (INCLUDE), to output audio-only headers. Choose Exclude (EXCLUDE) to remove the audio-only headers from your audio segments.

base_url: Option<String>

A partial URI prefix that will be prepended to each output in the media .m3u8 file. Can be used if base manifest is delivered from a different URL than the main .m3u8 file.

caption_language_mappings: Option<Vec<HlsCaptionLanguageMapping>>

Language to be used on Caption outputs

caption_language_setting: Option<HlsCaptionLanguageSetting>

Applies only to 608 Embedded output captions. Insert: Include CLOSED-CAPTIONS lines in the manifest. Specify at least one language in the CC1 Language Code field. One CLOSED-CAPTION line is added for each Language Code you specify. Make sure to specify the languages in the order in which they appear in the original source (if the source is embedded format) or the order of the caption selectors (if the source is other than embedded). Otherwise, languages in the manifest will not match up properly with the output captions. None: Include CLOSED-CAPTIONS=NONE line in the manifest. Omit: Omit any CLOSED-CAPTIONS line from the manifest.

client_cache: Option<HlsClientCache>

Disable this setting only when your workflow requires the #EXT-X-ALLOW-CACHE:no tag. Otherwise, keep the default value Enabled (ENABLED) and control caching in your video distribution set up. For example, use the Cache-Control http header.

codec_specification: Option<HlsCodecSpecification>

Specification to use (RFC-6381 or the default RFC-4281) during m3u8 playlist generation.

destination: Option<String>

Use Destination (Destination) to specify the S3 output location and the output filename base. Destination accepts format identifiers. If you do not specify the base filename in the URI, the service will use the filename of the input file. If your job has multiple inputs, the service uses the filename of the first input file.

destination_settings: Option<DestinationSettings>

Settings associated with the destination. Will vary based on the type of destination

directory_structure: Option<HlsDirectoryStructure>

Indicates whether segments should be placed in subdirectories.

encryption: Option<HlsEncryptionSettings>

DRM settings.

image_based_trick_play: Option<HlsImageBasedTrickPlay>

Specify whether MediaConvert generates images for trick play. Keep the default value, None (NONE), to not generate any images. Choose Thumbnail (THUMBNAIL) to generate tiled thumbnails. Choose Thumbnail and full frame (THUMBNAIL_AND_FULLFRAME) to generate tiled thumbnails and full-resolution images of single frames. MediaConvert creates a child manifest for each set of images that you generate and adds corresponding entries to the parent manifest. A common application for these images is Roku trick mode. The thumbnails and full-frame images that MediaConvert creates with this feature are compatible with this Roku specification: https://developer.roku.com/docs/developer-program/media-playback/trick-mode/hls-and-dash.md

manifest_compression: Option<HlsManifestCompression>

When set to GZIP, compresses HLS playlist.

manifest_duration_format: Option<HlsManifestDurationFormat>

Indicates whether the output manifest should use floating point values for segment duration.

min_final_segment_length: f64

Keep this setting at the default value of 0, unless you are troubleshooting a problem with how devices play back the end of your video asset. If you know that player devices are hanging on the final segment of your video because the length of your final segment is too short, use this setting to specify a minimum final segment length, in seconds. Choose a value that is greater than or equal to 1 and less than your segment length. When you specify a value for this setting, the encoder will combine any final segment that is shorter than the length that you specify with the previous segment. For example, your segment length is 3 seconds and your final segment is .5 seconds without a minimum final segment length; when you set the minimum final segment length to 1, your final segment is 3.5 seconds.

min_segment_length: i32

When set, Minimum Segment Size is enforced by looking ahead and back within the specified range for a nearby avail and extending the segment size if needed.

output_selection: Option<HlsOutputSelection>

Indicates whether the .m3u8 manifest file should be generated for this HLS output group.

program_date_time: Option<HlsProgramDateTime>

Includes or excludes EXT-X-PROGRAM-DATE-TIME tag in .m3u8 manifest files. The value is calculated as follows: either the program date and time are initialized using the input timecode source, or the time is initialized using the input timecode source and the date is initialized using the timestamp_offset.

program_date_time_period: i32

Period of insertion of EXT-X-PROGRAM-DATE-TIME entry, in seconds.

segment_control: Option<HlsSegmentControl>

When set to SINGLE_FILE, emits program as a single media resource (.ts) file, uses #EXT-X-BYTERANGE tags to index segment for playback.

segment_length: i32

Specify the length, in whole seconds, of each segment. When you don’t specify a value, MediaConvert defaults to 10. Related settings: Use Segment length control (SegmentLengthControl) to specify whether the encoder enforces this value strictly. Use Segment control (HlsSegmentControl) to specify whether MediaConvert creates separate segment files or one content file that has metadata to mark the segment boundaries.

segment_length_control: Option<HlsSegmentLengthControl>

Specify how you want MediaConvert to determine the segment length. Choose Exact (EXACT) to have the encoder use the exact length that you specify with the setting Segment length (SegmentLength). This might result in extra I-frames. Choose Multiple of GOP (GOP_MULTIPLE) to have the encoder round up the segment lengths to match the next GOP boundary.

segments_per_subdirectory: i32

Number of segments to write to a subdirectory before starting a new one. directoryStructure must be SINGLE_DIRECTORY for this setting to have an effect.

stream_inf_resolution: Option<HlsStreamInfResolution>

Include or exclude RESOLUTION attribute for video in EXT-X-STREAM-INF tag of variant manifest.

target_duration_compatibility_mode: Option<HlsTargetDurationCompatibilityMode>

When set to LEGACY, the segment target duration is always rounded up to the nearest integer value above its current value in seconds. When set to SPEC\_COMPLIANT, the segment target duration is rounded up to the nearest integer value if fraction seconds are greater than or equal to 0.5 (>= 0.5) and rounded down if less than 0.5 (< 0.5). You may need to use LEGACY if your client needs to ensure that the target duration is always longer than the actual duration of the segment. Some older players may experience interrupted playback when the actual duration of a track in a segment is longer than the target duration.

timed_metadata_id3_frame: Option<HlsTimedMetadataId3Frame>

Indicates ID3 frame that has the timecode.

timed_metadata_id3_period: i32

Timed Metadata interval in seconds.

timestamp_delta_milliseconds: i32

Provides an extra millisecond delta offset to fine tune the timestamps.

Implementations

impl HlsGroupSettings

pub fn builder() -> Builder

Creates a new builder-style object to manufacture HlsGroupSettings