Struct aws_sdk_mediaconvert::model::Job

#[non_exhaustive]
pub struct Job {

    pub acceleration_settings: Option<AccelerationSettings>,
    pub acceleration_status: Option<AccelerationStatus>,
    pub arn: Option<String>,
    pub billing_tags_source: Option<BillingTagsSource>,
    pub created_at: Option<Instant>,
    pub current_phase: Option<JobPhase>,
    pub error_code: i32,
    pub error_message: Option<String>,
    pub hop_destinations: Option<Vec<HopDestination>>,
    pub id: Option<String>,
    pub job_percent_complete: i32,
    pub job_template: Option<String>,
    pub messages: Option<JobMessages>,
    pub output_group_details: Option<Vec<OutputGroupDetail>>,
    pub priority: i32,
    pub queue: Option<String>,
    pub queue_transitions: Option<Vec<QueueTransition>>,
    pub retry_count: i32,
    pub role: Option<String>,
    pub settings: Option<JobSettings>,
    pub simulate_reserved_queue: Option<SimulateReservedQueue>,
    pub status: Option<JobStatus>,
    pub status_update_interval: Option<StatusUpdateInterval>,
    pub timing: Option<Timing>,
    pub user_metadata: Option<HashMap<String, String>>,

}

Each job converts an input file into an output file or files. For more information, see the User Guide at https://docs.aws.amazon.com/mediaconvert/latest/ug/what-is.html

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

acceleration_settings: Option<AccelerationSettings>

Accelerated transcoding can significantly speed up jobs with long, visually complex content.

acceleration_status: Option<AccelerationStatus>

Describes whether the current job is running with accelerated transcoding. For jobs that have Acceleration (AccelerationMode) set to DISABLED, AccelerationStatus is always NOT_APPLICABLE. For jobs that have Acceleration (AccelerationMode) set to ENABLED or PREFERRED, AccelerationStatus is one of the other states. AccelerationStatus is IN_PROGRESS initially, while the service determines whether the input files and job settings are compatible with accelerated transcoding. If they are, AcclerationStatus is ACCELERATED. If your input files and job settings aren’t compatible with accelerated transcoding, the service either fails your job or runs it without accelerated transcoding, depending on how you set Acceleration (AccelerationMode). When the service runs your job without accelerated transcoding, AccelerationStatus is NOT_ACCELERATED.

arn: Option<String>

An identifier for this resource that is unique within all of AWS.

billing_tags_source: Option<BillingTagsSource>

The tag type that AWS Billing and Cost Management will use to sort your AWS Elemental MediaConvert costs on any billing report that you set up.

created_at: Option<Instant>

The time, in Unix epoch format in seconds, when the job got created.

current_phase: Option<JobPhase>

A job’s phase can be PROBING, TRANSCODING OR UPLOADING

error_code: i32

Error code for the job

error_message: Option<String>

Error message of Job

hop_destinations: Option<Vec<HopDestination>>

Optional list of hop destinations.

id: Option<String>

A portion of the job’s ARN, unique within your AWS Elemental MediaConvert resources

job_percent_complete: i32

An estimate of how far your job has progressed. This estimate is shown as a percentage of the total time from when your job leaves its queue to when your output files appear in your output Amazon S3 bucket. AWS Elemental MediaConvert provides jobPercentComplete in CloudWatch STATUS_UPDATE events and in the response to GetJob and ListJobs requests. The jobPercentComplete estimate is reliable for the following input containers: Quicktime, Transport Stream, MP4, and MXF. For some jobs, the service can’t provide information about job progress. In those cases, jobPercentComplete returns a null value.

job_template: Option<String>

The job template that the job is created from, if it is created from a job template.

messages: Option<JobMessages>

Provides messages from the service about jobs that you have already successfully submitted.

output_group_details: Option<Vec<OutputGroupDetail>>

List of output group details

priority: i32

Relative priority on the job.

queue: Option<String>

When you create a job, you can specify a queue to send it to. If you don’t specify, the job will go to the default queue. For more about queues, see the User Guide topic at https://docs.aws.amazon.com/mediaconvert/latest/ug/what-is.html

queue_transitions: Option<Vec<QueueTransition>>

The job’s queue hopping history.

retry_count: i32

The number of times that the service automatically attempted to process your job after encountering an error.

role: Option<String>

The IAM role you use for creating this job. For details about permissions, see the User Guide topic at the User Guide at https://docs.aws.amazon.com/mediaconvert/latest/ug/iam-role.html

settings: Option<JobSettings>

JobSettings contains all the transcode settings for a job.

simulate_reserved_queue: Option<SimulateReservedQueue>

Enable this setting when you run a test job to estimate how many reserved transcoding slots (RTS) you need. When this is enabled, MediaConvert runs your job from an on-demand queue with similar performance to what you will see with one RTS in a reserved queue. This setting is disabled by default.

status: Option<JobStatus>

A job’s status can be SUBMITTED, PROGRESSING, COMPLETE, CANCELED, or ERROR.

status_update_interval: Option<StatusUpdateInterval>

Specify how often MediaConvert sends STATUS_UPDATE events to Amazon CloudWatch Events. Set the interval, in seconds, between status updates. MediaConvert sends an update at this interval from the time the service begins processing your job to the time it completes the transcode or encounters an error.

timing: Option<Timing>

Information about when jobs are submitted, started, and finished is specified in Unix epoch format in seconds.

user_metadata: Option<HashMap<String, String>>

User-defined metadata that you want to associate with an MediaConvert job. You specify metadata in key/value pairs.

Implementations

impl Job

pub fn builder() -> Builder

Creates a new builder-style object to manufacture Job