Struct aws_sdk_mediaconvert::model::JobTemplateSettings

#[non_exhaustive]
pub struct JobTemplateSettings {
    pub ad_avail_offset: i32,
    pub avail_blanking: Option<AvailBlanking>,
    pub esam: Option<EsamSettings>,
    pub extended_data_services: Option<ExtendedDataServices>,
    pub inputs: Option<Vec<InputTemplate>>,
    pub kantar_watermark: Option<KantarWatermarkSettings>,
    pub motion_image_inserter: Option<MotionImageInserter>,
    pub nielsen_configuration: Option<NielsenConfiguration>,
    pub nielsen_non_linear_watermark: Option<NielsenNonLinearWatermarkSettings>,
    pub output_groups: Option<Vec<OutputGroup>>,
    pub timecode_config: Option<TimecodeConfig>,
    pub timed_metadata_insertion: Option<TimedMetadataInsertion>,
}

JobTemplateSettings contains all the transcode settings saved in the template that will be applied to jobs created from it.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

ad_avail_offset: i32

When specified, this offset (in milliseconds) is added to the input Ad Avail PTS time.

avail_blanking: Option<AvailBlanking>

Settings for ad avail blanking. Video can be blanked or overlaid with an image, and audio muted during SCTE-35 triggered ad avails.

esam: Option<EsamSettings>

Settings for Event Signaling And Messaging (ESAM). If you don’t do ad insertion, you can ignore these settings.

extended_data_services: Option<ExtendedDataServices>

Hexadecimal value as per EIA-608 Line 21 Data Services, section 9.5.1.5 05h Content Advisory.

inputs: Option<Vec<InputTemplate>>

Use Inputs (inputs) to define the source file used in the transcode job. There can only be one input in a job template. Using the API, you can include multiple inputs when referencing a job template.

kantar_watermark: Option<KantarWatermarkSettings>

Use these settings only when you use Kantar watermarking. Specify the values that MediaConvert uses to generate and place Kantar watermarks in your output audio. These settings apply to every output in your job. In addition to specifying these values, you also need to store your Kantar credentials in AWS Secrets Manager. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/kantar-watermarking.html.

motion_image_inserter: Option<MotionImageInserter>

Overlay motion graphics on top of your video. The motion graphics that you specify here appear on all outputs in all output groups. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/motion-graphic-overlay.html.

nielsen_configuration: Option<NielsenConfiguration>

Settings for your Nielsen configuration. If you don’t do Nielsen measurement and analytics, ignore these settings. When you enable Nielsen configuration (nielsenConfiguration), MediaConvert enables PCM to ID3 tagging for all outputs in the job. To enable Nielsen configuration programmatically, include an instance of nielsenConfiguration in your JSON job specification. Even if you don’t include any children of nielsenConfiguration, you still enable the setting.

nielsen_non_linear_watermark: Option<NielsenNonLinearWatermarkSettings>

Ignore these settings unless you are using Nielsen non-linear watermarking. Specify the values that MediaConvert uses to generate and place Nielsen watermarks in your output audio. In addition to specifying these values, you also need to set up your cloud TIC server. These settings apply to every output in your job. The MediaConvert implementation is currently with the following Nielsen versions: Nielsen Watermark SDK Version 5.2.1 Nielsen NLM Watermark Engine Version 1.2.7 Nielsen Watermark Authenticator [SID_TIC] Version [5.0.0]

output_groups: Option<Vec<OutputGroup>>

(OutputGroups) contains one group of settings for each set of outputs that share a common package type. All unpackaged files (MPEG-4, MPEG-2 TS, Quicktime, MXF, and no container) are grouped in a single output group as well. Required in (OutputGroups) is a group of settings that apply to the whole group. This required object depends on the value you set for (Type) under (OutputGroups)>(OutputGroupSettings). Type, settings object pairs are as follows. * FILE_GROUP_SETTINGS, FileGroupSettings * HLS_GROUP_SETTINGS, HlsGroupSettings * DASH_ISO_GROUP_SETTINGS, DashIsoGroupSettings * MS_SMOOTH_GROUP_SETTINGS, MsSmoothGroupSettings * CMAF_GROUP_SETTINGS, CmafGroupSettings

timecode_config: Option<TimecodeConfig>

These settings control how the service handles timecodes throughout the job. These settings don’t affect input clipping.

timed_metadata_insertion: Option<TimedMetadataInsertion>

Enable Timed metadata insertion (TimedMetadataInsertion) to include ID3 tags in any HLS outputs. To include timed metadata, you must enable it here, enable it in each output container, and specify tags and timecodes in ID3 insertion (Id3Insertion) objects.

Implementations

impl JobTemplateSettings

pub fn builder() -> Builder

Creates a new builder-style object to manufacture JobTemplateSettings