Struct aws_sdk_mediaconvert::model::M2tsSettings

#[non_exhaustive]
pub struct M2tsSettings {

    pub audio_buffer_model: Option<M2tsAudioBufferModel>,
    pub audio_duration: Option<M2tsAudioDuration>,
    pub audio_frames_per_pes: i32,
    pub audio_pids: Option<Vec<i32>>,
    pub bitrate: i32,
    pub buffer_model: Option<M2tsBufferModel>,
    pub data_pts_control: Option<M2tsDataPtsControl>,
    pub dvb_nit_settings: Option<DvbNitSettings>,
    pub dvb_sdt_settings: Option<DvbSdtSettings>,
    pub dvb_sub_pids: Option<Vec<i32>>,
    pub dvb_tdt_settings: Option<DvbTdtSettings>,
    pub dvb_teletext_pid: i32,
    pub ebp_audio_interval: Option<M2tsEbpAudioInterval>,
    pub ebp_placement: Option<M2tsEbpPlacement>,
    pub es_rate_in_pes: Option<M2tsEsRateInPes>,
    pub force_ts_video_ebp_order: Option<M2tsForceTsVideoEbpOrder>,
    pub fragment_time: f64,
    pub max_pcr_interval: i32,
    pub min_ebp_interval: i32,
    pub nielsen_id3: Option<M2tsNielsenId3>,
    pub null_packet_bitrate: f64,
    pub pat_interval: i32,
    pub pcr_control: Option<M2tsPcrControl>,
    pub pcr_pid: i32,
    pub pmt_interval: i32,
    pub pmt_pid: i32,
    pub private_metadata_pid: i32,
    pub program_number: i32,
    pub rate_mode: Option<M2tsRateMode>,
    pub scte35_esam: Option<M2tsScte35Esam>,
    pub scte35_pid: i32,
    pub scte35_source: Option<M2tsScte35Source>,
    pub segmentation_markers: Option<M2tsSegmentationMarkers>,
    pub segmentation_style: Option<M2tsSegmentationStyle>,
    pub segmentation_time: f64,
    pub timed_metadata_pid: i32,
    pub transport_stream_id: i32,
    pub video_pid: i32,

}

MPEG-2 TS container settings. These apply to outputs in a File output group when the output’s container (ContainerType) is MPEG-2 Transport Stream (M2TS). In these assets, data is organized by the program map table (PMT). Each transport stream program contains subsets of data, including audio, video, and metadata. Each of these subsets of data has a numerical label called a packet identifier (PID). Each transport stream program corresponds to one MediaConvert output. The PMT lists the types of data in a program along with their PID. Downstream systems and players use the program map table to look up the PID for each type of data it accesses and then uses the PIDs to locate specific data within the asset.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

audio_buffer_model: Option<M2tsAudioBufferModel>

Selects between the DVB and ATSC buffer models for Dolby Digital audio.

audio_duration: Option<M2tsAudioDuration>

Specify this setting only when your output will be consumed by a downstream repackaging workflow that is sensitive to very small duration differences between video and audio. For this situation, choose Match video duration (MATCH_VIDEO_DURATION). In all other cases, keep the default value, Default codec duration (DEFAULT_CODEC_DURATION). When you choose Match video duration, MediaConvert pads the output audio streams with silence or trims them to ensure that the total duration of each audio stream is at least as long as the total duration of the video stream. After padding or trimming, the audio stream duration is no more than one frame longer than the video stream. MediaConvert applies audio padding or trimming only to the end of the last segment of the output. For unsegmented outputs, MediaConvert adds padding only to the end of the file. When you keep the default value, any minor discrepancies between audio and video duration will depend on your output audio codec.

audio_frames_per_pes: i32

The number of audio frames to insert for each PES packet.

audio_pids: Option<Vec<i32>>

Specify the packet identifiers (PIDs) for any elementary audio streams you include in this output. Specify multiple PIDs as a JSON array. Default is the range 482-492.

bitrate: i32

Specify the output bitrate of the transport stream in bits per second. Setting to 0 lets the muxer automatically determine the appropriate bitrate. Other common values are 3750000, 7500000, and 15000000.

buffer_model: Option<M2tsBufferModel>

Controls what buffer model to use for accurate interleaving. If set to MULTIPLEX, use multiplex buffer model. If set to NONE, this can lead to lower latency, but low-memory devices may not be able to play back the stream without interruptions.

data_pts_control: Option<M2tsDataPtsControl>

If you select ALIGN_TO_VIDEO, MediaConvert writes captions and data packets with Presentation Timestamp (PTS) values greater than or equal to the first video packet PTS (MediaConvert drops captions and data packets with lesser PTS values). Keep the default value (AUTO) to allow all PTS values.

dvb_nit_settings: Option<DvbNitSettings>

Use these settings to insert a DVB Network Information Table (NIT) in the transport stream of this output. When you work directly in your JSON job specification, include this object only when your job has a transport stream output and the container settings contain the object M2tsSettings.

dvb_sdt_settings: Option<DvbSdtSettings>

Use these settings to insert a DVB Service Description Table (SDT) in the transport stream of this output. When you work directly in your JSON job specification, include this object only when your job has a transport stream output and the container settings contain the object M2tsSettings.

dvb_sub_pids: Option<Vec<i32>>

Specify the packet identifiers (PIDs) for DVB subtitle data included in this output. Specify multiple PIDs as a JSON array. Default is the range 460-479.

dvb_tdt_settings: Option<DvbTdtSettings>

Use these settings to insert a DVB Time and Date Table (TDT) in the transport stream of this output. When you work directly in your JSON job specification, include this object only when your job has a transport stream output and the container settings contain the object M2tsSettings.

dvb_teletext_pid: i32

Specify the packet identifier (PID) for DVB teletext data you include in this output. Default is 499.

ebp_audio_interval: Option<M2tsEbpAudioInterval>

When set to VIDEO_AND_FIXED_INTERVALS, audio EBP markers will be added to partitions 3 and 4. The interval between these additional markers will be fixed, and will be slightly shorter than the video EBP marker interval. When set to VIDEO_INTERVAL, these additional markers will not be inserted. Only applicable when EBP segmentation markers are is selected (segmentationMarkers is EBP or EBP_LEGACY).

ebp_placement: Option<M2tsEbpPlacement>

Selects which PIDs to place EBP markers on. They can either be placed only on the video PID, or on both the video PID and all audio PIDs. Only applicable when EBP segmentation markers are is selected (segmentationMarkers is EBP or EBP_LEGACY).

es_rate_in_pes: Option<M2tsEsRateInPes>

Controls whether to include the ES Rate field in the PES header.

force_ts_video_ebp_order: Option<M2tsForceTsVideoEbpOrder>

Keep the default value (DEFAULT) unless you know that your audio EBP markers are incorrectly appearing before your video EBP markers. To correct this problem, set this value to Force (FORCE).

fragment_time: f64

The length, in seconds, of each fragment. Only used with EBP markers.

max_pcr_interval: i32

Specify the maximum time, in milliseconds, between Program Clock References (PCRs) inserted into the transport stream.

min_ebp_interval: i32

When set, enforces that Encoder Boundary Points do not come within the specified time interval of each other by looking ahead at input video. If another EBP is going to come in within the specified time interval, the current EBP is not emitted, and the segment is “stretched” to the next marker. The lookahead value does not add latency to the system. The Live Event must be configured elsewhere to create sufficient latency to make the lookahead accurate.

nielsen_id3: Option<M2tsNielsenId3>

If INSERT, Nielsen inaudible tones for media tracking will be detected in the input audio and an equivalent ID3 tag will be inserted in the output.

null_packet_bitrate: f64

Value in bits per second of extra null packets to insert into the transport stream. This can be used if a downstream encryption system requires periodic null packets.

pat_interval: i32

The number of milliseconds between instances of this table in the output transport stream.

pcr_control: Option<M2tsPcrControl>

When set to PCR_EVERY_PES_PACKET, a Program Clock Reference value is inserted for every Packetized Elementary Stream (PES) header. This is effective only when the PCR PID is the same as the video or audio elementary stream.

pcr_pid: i32

Specify the packet identifier (PID) for the program clock reference (PCR) in this output. If you do not specify a value, the service will use the value for Video PID (VideoPid).

pmt_interval: i32

Specify the number of milliseconds between instances of the program map table (PMT) in the output transport stream.

pmt_pid: i32

Specify the packet identifier (PID) for the program map table (PMT) itself. Default is 480.

private_metadata_pid: i32

Specify the packet identifier (PID) of the private metadata stream. Default is 503.

program_number: i32

Use Program number (programNumber) to specify the program number used in the program map table (PMT) for this output. Default is 1. Program numbers and program map tables are parts of MPEG-2 transport stream containers, used for organizing data.

rate_mode: Option<M2tsRateMode>

When set to CBR, inserts null packets into transport stream to fill specified bitrate. When set to VBR, the bitrate setting acts as the maximum bitrate, but the output will not be padded up to that bitrate.

scte35_esam: Option<M2tsScte35Esam>

Include this in your job settings to put SCTE-35 markers in your HLS and transport stream outputs at the insertion points that you specify in an ESAM XML document. Provide the document in the setting SCC XML (sccXml).

scte35_pid: i32

Specify the packet identifier (PID) of the SCTE-35 stream in the transport stream.

scte35_source: Option<M2tsScte35Source>

For SCTE-35 markers from your input– Choose Passthrough (PASSTHROUGH) if you want SCTE-35 markers that appear in your input to also appear in this output. Choose None (NONE) if you don’t want SCTE-35 markers in this output. For SCTE-35 markers from an ESAM XML document– Choose None (NONE). Also provide the ESAM XML as a string in the setting Signal processing notification XML (sccXml). Also enable ESAM SCTE-35 (include the property scte35Esam).

segmentation_markers: Option<M2tsSegmentationMarkers>

Inserts segmentation markers at each segmentation_time period. rai_segstart sets the Random Access Indicator bit in the adaptation field. rai_adapt sets the RAI bit and adds the current timecode in the private data bytes. psi_segstart inserts PAT and PMT tables at the start of segments. ebp adds Encoder Boundary Point information to the adaptation field as per OpenCable specification OC-SP-EBP-I01-130118. ebp_legacy adds Encoder Boundary Point information to the adaptation field using a legacy proprietary format.

segmentation_style: Option<M2tsSegmentationStyle>

The segmentation style parameter controls how segmentation markers are inserted into the transport stream. With avails, it is possible that segments may be truncated, which can influence where future segmentation markers are inserted. When a segmentation style of “reset_cadence” is selected and a segment is truncated due to an avail, we will reset the segmentation cadence. This means the subsequent segment will have a duration of of $segmentation_time seconds. When a segmentation style of “maintain_cadence” is selected and a segment is truncated due to an avail, we will not reset the segmentation cadence. This means the subsequent segment will likely be truncated as well. However, all segments after that will have a duration of $segmentation_time seconds. Note that EBP lookahead is a slight exception to this rule.

segmentation_time: f64

Specify the length, in seconds, of each segment. Required unless markers is set to none.

timed_metadata_pid: i32

Specify the packet identifier (PID) for timed metadata in this output. Default is 502.

transport_stream_id: i32

Specify the ID for the transport stream itself in the program map table for this output. Transport stream IDs and program map tables are parts of MPEG-2 transport stream containers, used for organizing data.

video_pid: i32

Specify the packet identifier (PID) of the elementary video stream in the transport stream.

Implementations

impl M2tsSettings

pub fn builder() -> Builder

Creates a new builder-style object to manufacture M2tsSettings