Struct aws_sdk_mediaconvert::model::Vp9Settings

#[non_exhaustive]
pub struct Vp9Settings {

    pub bitrate: i32,
    pub framerate_control: Option<Vp9FramerateControl>,
    pub framerate_conversion_algorithm: Option<Vp9FramerateConversionAlgorithm>,
    pub framerate_denominator: i32,
    pub framerate_numerator: i32,
    pub gop_size: f64,
    pub hrd_buffer_size: i32,
    pub max_bitrate: i32,
    pub par_control: Option<Vp9ParControl>,
    pub par_denominator: i32,
    pub par_numerator: i32,
    pub quality_tuning_level: Option<Vp9QualityTuningLevel>,
    pub rate_control_mode: Option<Vp9RateControlMode>,

}

Required when you set (Codec) under (VideoDescription)>(CodecSettings) to the value VP9.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

bitrate: i32

Target bitrate in bits/second. For example, enter five megabits per second as 5000000.

framerate_control: Option<Vp9FramerateControl>

If you are using the console, use the Framerate setting to specify the frame rate for this output. If you want to keep the same frame rate as the input video, choose Follow source. If you want to do frame rate conversion, choose a frame rate from the dropdown list or choose Custom. The framerates shown in the dropdown list are decimal approximations of fractions. If you choose Custom, specify your frame rate as a fraction. If you are creating your transcoding job specification as a JSON file without the console, use FramerateControl to specify which value the service uses for the frame rate for this output. Choose INITIALIZE_FROM_SOURCE if you want the service to use the frame rate from the input. Choose SPECIFIED if you want the service to use the frame rate you specify in the settings FramerateNumerator and FramerateDenominator.

framerate_conversion_algorithm: Option<Vp9FramerateConversionAlgorithm>

Choose the method that you want MediaConvert to use when increasing or decreasing the frame rate. We recommend using drop duplicate (DUPLICATE_DROP) for numerically simple conversions, such as 60 fps to 30 fps. For numerically complex conversions, you can use interpolate (INTERPOLATE) to avoid stutter. This results in a smooth picture, but might introduce undesirable video artifacts. For complex frame rate conversions, especially if your source video has already been converted from its original cadence, use FrameFormer (FRAMEFORMER) to do motion-compensated interpolation. FrameFormer chooses the best conversion method frame by frame. Note that using FrameFormer increases the transcoding time and incurs a significant add-on cost.

framerate_denominator: i32

When you use the API for transcode jobs that use frame rate conversion, specify the frame rate as a fraction. For example, 24000 / 1001 = 23.976 fps. Use FramerateDenominator to specify the denominator of this fraction. In this example, use 1001 for the value of FramerateDenominator. When you use the console for transcode jobs that use frame rate conversion, provide the value as a decimal number for Framerate. In this example, specify 23.976.

framerate_numerator: i32

When you use the API for transcode jobs that use frame rate conversion, specify the frame rate as a fraction. For example, 24000 / 1001 = 23.976 fps. Use FramerateNumerator to specify the numerator of this fraction. In this example, use 24000 for the value of FramerateNumerator. When you use the console for transcode jobs that use frame rate conversion, provide the value as a decimal number for Framerate. In this example, specify 23.976.

gop_size: f64

GOP Length (keyframe interval) in frames. Must be greater than zero.

hrd_buffer_size: i32

Size of buffer (HRD buffer model) in bits. For example, enter five megabits as 5000000.

max_bitrate: i32

Ignore this setting unless you set qualityTuningLevel to MULTI_PASS. Optional. Specify the maximum bitrate in bits/second. For example, enter five megabits per second as 5000000. The default behavior uses twice the target bitrate as the maximum bitrate.

par_control: Option<Vp9ParControl>

Optional. Specify how the service determines the pixel aspect ratio for this output. The default behavior is to use the same pixel aspect ratio as your input video.

par_denominator: i32

Required when you set Pixel aspect ratio (parControl) to SPECIFIED. On the console, this corresponds to any value other than Follow source. When you specify an output pixel aspect ratio (PAR) that is different from your input video PAR, provide your output PAR as a ratio. For example, for D1/DV NTSC widescreen, you would specify the ratio 40:33. In this example, the value for parDenominator is 33.

par_numerator: i32

Required when you set Pixel aspect ratio (parControl) to SPECIFIED. On the console, this corresponds to any value other than Follow source. When you specify an output pixel aspect ratio (PAR) that is different from your input video PAR, provide your output PAR as a ratio. For example, for D1/DV NTSC widescreen, you would specify the ratio 40:33. In this example, the value for parNumerator is 40.

quality_tuning_level: Option<Vp9QualityTuningLevel>

Optional. Use Quality tuning level (qualityTuningLevel) to choose how you want to trade off encoding speed for output video quality. The default behavior is faster, lower quality, multi-pass encoding.

rate_control_mode: Option<Vp9RateControlMode>

With the VP9 codec, you can use only the variable bitrate (VBR) rate control mode.

Implementations

impl Vp9Settings

pub fn builder() -> Builder

Creates a new builder-style object to manufacture Vp9Settings